A third of main.rs was one test module
`main.rs` was 11,297 lines, and 3,534 of them were a single `mod tests`
holding 120 tests about everything the shell does -- windows, quitting,
navigation, updates, diagnostics -- in one undifferentiated block.

It is a directory now, grouped the way the code it covers is grouped:
splash and window placement, quitting and its prompt, navigation,
updates and install, runtime, locald, diagnostics, agent host, config.
Every test moved verbatim and all 165 still pass, which is the check that
the move was a move.

Shared helpers were the one thing that could not simply be sorted: five
of them are used from several groups, so they live in `tests/mod.rs`
where every group can see them rather than being duplicated per file.

The relative includes needed rewriting -- these files sit a directory
deeper than `main.rs`, so `include_str!("../tauri.conf.json")` no longer
pointed at anything. That is the kind of thing a split breaks silently if
the tests are not run, which is why the count above matters more than the
line counts.

main.rs is 7,765 lines now. Still far past DES-09's 600; the source
modules are the next pass.
