+++ desktop/scripts/file-size-baseline.json
@@ -18,5 +18,5 @@
   "desktop/locald/src/operator_config.rs": 2423,
   "desktop/locald/src/sharing.rs": 2254,
   "desktop/src/artifact_install.rs": 2272,
-  "desktop/src/main.rs": 11475
+  "desktop/src/main.rs": 7850
 }
