+++ desktop/src/main.rs
@@ -7846,3630 +7846,5 @@ fn main() {
 }
 
 #[cfg(test)]
-mod tests {
-    use std::fs::File;
-
-    /// Cloud and local must never share a store, and neither may drift between
-    /// launches.
-    ///
-    /// The identifiers are constants precisely so this can be asserted. If one
-    /// were ever derived from something per-run, a restart would look like a
-    /// different server and silently sign the user out of the one they kept.
-    #[cfg(target_os = "macos")]
-    #[test]
-    fn each_server_gets_its_own_session_store() {
-        let hosted = super::session_partition_id("hosted");
-        let local = super::session_partition_id("local");
-        assert_ne!(hosted, local);
-        // Stable across calls, which is what makes a session survive a restart.
-        assert_eq!(hosted, super::session_partition_id("hosted"));
-        assert_eq!(local, super::session_partition_id("local"));
-        // Anything that is not the hosted server is the local one. "undecided"
-        // reaches here on a first launch that has not been answered yet, and it
-        // must not land in the cloud store.
-        assert_eq!(local, super::session_partition_id("undecided"));
-        assert_eq!(local, super::session_partition_id(""));
-    }
-
-    /// The shell and the daemon must derive the same endpoint name.
-    ///
-    /// This exact assertion is duplicated in locald/src/paths.rs, over the same
-    /// path and the same expected string, because the code that produces it is
-    /// duplicated too and cannot cheaply be shared -- locald is a sidecar
-    /// binary, not a library this crate links. They drifted once: this side
-    /// hashed the root unnormalised while the daemon lowercased it, so on every
-    /// default Windows install the app opened a pipe its own daemon never
-    /// listened on and called it "control endpoint unavailable". Changing this
-    /// value without changing the other one is the bug.
-    #[cfg(windows)]
-    #[test]
-    fn named_pipe_name_matches_the_one_locald_listens_on() {
-        assert_eq!(
-            super::locald_pipe_name(std::path::Path::new(
-                r"C:\Users\Example\AppData\Local\Lemma\locald"
-            )),
-            r"LOCAL\work.lemma.locald.a5c86f3cbfe10caf"
-        );
-        // Every spelling of one directory is one endpoint.
-        assert_eq!(
-            super::locald_pipe_name(std::path::Path::new(
-                r"C:\Users\Example\AppData\Local\Lemma\locald"
-            )),
-            super::locald_pipe_name(std::path::Path::new(
-                r"c:/users/example/appdata/local/lemma/locald/"
-            ))
-        );
-    }
-
-    #[test]
-    fn a_menu_verb_runs_once() {
-        // `app.on_menu_event` receives events from every menu the app owns, the
-        // tray's included. The tray builder registered a second handler on top,
-        // so every tray verb ran twice: two confirmation dialogs stacked on each
-        // other, two stops, two restarts.
-        // The needle is assembled at compile time so it never appears whole in
-        // this file -- a source-scanning test that spells out what it is looking
-        // for finds itself, which is how the first two versions of this failed
-        // on the fix they were written to protect.
-        let needle = concat!("on_menu_event", "(|app, event|");
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        assert_eq!(
-            source.matches(needle).count(),
-            1,
-            "exactly one menu event handler, or every verb fires once per handler"
-        );
-    }
-
-    #[test]
-    fn the_splash_draws_something_for_every_state() {
-        // `renderState` is the only thing that paints the splash and the only
-        // thing that schedules the move to the workspace. It used to return
-        // early for a state with no phase, and again for an undecided
-        // connection mode -- so those states left the bare static logo on
-        // screen indefinitely, which is what a stuck first run looked like.
-        let splash = include_str!("../ui/index.html").replace("\r\n", "\n");
-        let body = {
-            let start = splash
-                .find("function renderState(s) {")
-                .expect("renderState exists");
-            let end = splash[start..]
-                .find("\n  function ")
-                .map_or(splash.len(), |offset| start + offset);
-            &splash[start..end]
-        };
-        assert!(
-            !body.contains("if (!s || !s.phaseKey) return;"),
-            "a state without a phase must still render"
-        );
-        assert!(
-            !body.contains(r#"if (s.mode === "undecided") return;"#),
-            "an undecided mode must show the chooser, not an empty screen"
-        );
-        assert!(
-            body.contains("showChooser();"),
-            "the undecided branch must offer the connection choice"
-        );
-    }
-
-    #[test]
-    fn no_lock_is_held_across_the_runtime_install() {
-        // `ensure_locald` used to take `locald_connect` and then install the
-        // runtime -- hundreds of megabytes -- while holding it. Local settings'
-        // heartbeat takes the same path, so opening settings during a first
-        // install blocked for the whole install. The install has its own
-        // single-flight now, and it must come first.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let body = function_body(&source, "fn ensure_locald(app: &AppHandle)");
-        let install = body
-            .find("ensure_runtime_artifacts(app)")
-            .expect("ensure_locald installs the runtime");
-        let connect = body
-            .find("locald_connect.lock()")
-            .expect("ensure_locald takes the connect guard");
-        assert!(
-            install < connect,
-            "the runtime install must happen before the connect guard is taken, \
-             not inside it"
-        );
-        assert!(
-            body.contains("runtime_install.lock()"),
-            "the install still needs its own single-flight so two callers cannot \
-             download at once"
-        );
-    }
-
-    /// Uninstalling asked to delete Lemma's data and deleted nothing.
-    ///
-    /// Tauri's uninstaller offers a "delete application data" checkbox, and
-    /// what it removes is `%APPDATA%\<identifier>` and
-    /// `%LOCALAPPDATA%\<identifier>` -- work.lemma.desktop. Lemma's data is in
-    /// `%LOCALAPPDATA%\Lemma`, and most of it is not files at all: it is a
-    /// registered WSL distribution whose ext4.vhdx holds every workspace,
-    /// database and image. So the box removed nothing anybody had, and an
-    /// uninstall left several gigabytes and a registered distribution that
-    /// only `wsl --unregister` could clear.
-    ///
-    /// The hook is what closes that, and these are the three things about it
-    /// that have to stay true. It cannot be executed from here -- NSIS runs
-    /// only on Windows, and only during a real uninstall -- so CI's Windows
-    /// job building the installer is what proves it parses.
-    /// Clicking the Dock icon with no window open did nothing at all.
-    ///
-    /// On macOS closing the last window leaves the app running, so this is the
-    /// one gesture whose whole purpose is bringing it back -- and the arm that
-    /// handles it only ever showed a window that already existed. This is the
-    /// part of the fix with cases in it; the arm itself needs a Tauri runtime.
-    #[cfg(target_os = "macos")]
-    #[test]
-    fn reopening_with_every_window_closed_goes_somewhere() {
-        assert_eq!(
-            reopen_target("local", true, false, "http://app.127.0.0.1.sslip.io:1/"),
-            ReopenTarget::Workspace("http://app.127.0.0.1.sslip.io:1/".into()),
-            "a running local stack goes straight back to the workspace"
-        );
-
-        // Hosted has no local stack to be ready for.
-        assert_eq!(
-            reopen_target("hosted", false, false, ""),
-            ReopenTarget::Hosted
-        );
-
-        for (label, ready, error, url) in [
-            ("still starting", false, false, ""),
-            ("failed", true, true, "http://app.127.0.0.1.sslip.io:1/"),
-            ("ready but with no url yet", true, false, ""),
-            (
-                "undecided mode",
-                true,
-                false,
-                "http://app.127.0.0.1.sslip.io:1/",
-            ),
-        ] {
-            let mode = if label == "undecided mode" {
-                "undecided"
-            } else {
-                "local"
-            };
-            assert_eq!(
-                reopen_target(mode, ready, error, url),
-                ReopenTarget::Splash,
-                "{label}: the splash is where the state and the actions are"
-            );
-        }
-    }
-
-    /// Which operation's progress the splash shows, in every case.
-    ///
-    /// The daemon serves one operation at a time, but several surfaces ask and
-    /// their replies interleave, so an event carries the operation it belongs
-    /// to and this decides whether it is the one on screen. It had no test at
-    /// all: `handle_locald_event` needs an `AppHandle` and a Tauri runtime, so
-    /// the only way to exercise any of this was to use the app and watch.
-    ///
-    /// Showing the wrong one is not cosmetic. Its phases and its failures
-    /// describe work the person watching did not start.
-    #[test]
-    fn only_the_operation_on_screen_moves_the_splash() {
-        let finished = vec!["install-1".to_string(), "start-1".to_string()];
-
-        // Nothing on screen: the first event of an unfinished operation takes it.
-        assert_eq!(
-            admit_locald_event("", &finished, "start-2"),
-            EventAdmission::Adopt
-        );
-
-        // Nothing on screen, but this one already ran to completion. A late
-        // straggler must not reopen a finished run's progress.
-        assert_eq!(
-            admit_locald_event("", &finished, "start-1"),
-            EventAdmission::Ignore
-        );
-
-        // The one being shown.
-        assert_eq!(
-            admit_locald_event("start-2", &finished, "start-2"),
-            EventAdmission::Apply
-        );
-
-        // Another surface's, while one is on screen.
-        assert_eq!(
-            admit_locald_event("start-2", &finished, "install-9"),
-            EventAdmission::Ignore
-        );
-
-        // Completion is only consulted when nothing is active: an operation
-        // that is on screen keeps its own events even if a stale entry for it
-        // survives in the ring.
-        assert_eq!(
-            admit_locald_event("start-1", &finished, "start-1"),
-            EventAdmission::Apply
-        );
-    }
-
-    #[test]
-    fn the_windows_uninstaller_removes_the_data_the_checkbox_promises() {
-        // Normalised: CI's Windows runner checks the tree out with CRLF, and a
-        // needle spanning a line break would find nothing there.
-        let config = include_str!("../tauri.windows.conf.json").replace("\r\n", "\n");
-        let hooks = include_str!("../installer/hooks.nsh").replace("\r\n", "\n");
-
-        assert!(
-            config.contains(r#""installerHooks": "installer/hooks.nsh""#),
-            "the uninstaller's data checkbox does nothing without this hook"
-        );
-        // Pinned rather than left to Tauri's default, which decides whether
-        // the app lands in Program Files and needs Administrator.
-        assert!(
-            config.contains(r#""installMode": "currentUser""#),
-            "the install mode has to be a decision here, not a default \
-             elsewhere that can change under us"
-        );
-
-        assert!(
-            hooks.contains("!macro NSIS_HOOK_PREUNINSTALL"),
-            "PRE, not POST: by POSTUNINSTALL lemma-locald.exe has already been \
-             deleted, and it is the thing that knows how to unregister the guest"
-        );
-        assert!(
-            !hooks.contains("NSIS_HOOK_POSTUNINSTALL"),
-            "nothing can run from $INSTDIR after the files are gone"
-        );
-        assert!(
-            hooks.contains("$DeleteAppDataCheckboxState = 1") && hooks.contains("$UpdateMode <> 1"),
-            "erasing local data must happen only when it was asked for, and \
-             never during an update"
-        );
-        let purge = hooks
-            .find("reset --confirm=erase-local-lemma")
-            .expect("the hook runs locald's own reset rather than reimplementing it");
-        let stop = hooks
-            .find("taskkill /F /T /IM lemma-locald.exe")
-            .expect("reset refuses while the daemon is answering, so it stops first");
-        assert!(
-            stop < purge,
-            "the daemon has to be stopped before the reset, or the reset refuses"
-        );
-    }
-
-    /// On Windows the updater exits this process to run the installer, so
-    /// there is no line after `install` in which to record anything. An
-    /// installer cancelled at the UAC prompt, or interrupted by a reboot, left
-    /// the app running the old version with nothing anywhere saying an update
-    /// had been attempted -- so the next launch, and the user, had no account
-    /// of why they were still on the version they were on.
-    #[test]
-    fn an_update_attempt_is_read_against_the_version_actually_running() {
-        let attempt = json!({"schema_version": 1, "from": "0.7.2", "to": "0.7.3"});
-
-        assert_eq!(
-            classify_update_attempt(Some(&attempt), "0.7.3"),
-            UpdateAttempt::Landed { to: "0.7.3".into() },
-            "running the version it aimed at is the update having happened"
-        );
-        assert_eq!(
-            classify_update_attempt(Some(&attempt), "0.7.2"),
-            UpdateAttempt::DidNotLand { to: "0.7.3".into() },
-            "still on the version it started from means the installer never ran"
-        );
-        // Neither end matches: a third version got installed in between, and
-        // this record explains nothing about the launch reading it.
-        assert_eq!(
-            classify_update_attempt(Some(&attempt), "0.8.0"),
-            UpdateAttempt::Unexplained
-        );
-        assert_eq!(
-            classify_update_attempt(Some(&json!({"schema_version": 1})), "0.7.2"),
-            UpdateAttempt::Unexplained,
-            "a record missing its versions cannot be acted on"
-        );
-        assert_eq!(classify_update_attempt(None, "0.7.2"), UpdateAttempt::None);
-    }
-
-    /// The record lives where an install cannot take it.
-    ///
-    /// Measured on Windows rather than assumed: the NSIS installer puts Lemma
-    /// in `%LOCALAPPDATA%\Lemma`, which is also where `locald` keeps its
-    /// state, so "beside the app" and "in the state root" are the same
-    /// directory there. What separates them is that the uninstaller removes
-    /// only the files it installed -- verified by leaving a file in
-    /// `locald\` across an `uninstall.exe /S` and finding it still there --
-    /// so anything under `locald` survives an install, an update and an
-    /// uninstall alike, and the shipped executables do not.
-    ///
-    /// Which makes the state root the only correct home for a record whose
-    /// whole job is to outlive the installer.
-    #[test]
-    fn the_update_record_is_kept_where_an_install_cannot_take_it() {
-        let path = update_attempt_path();
-        assert!(
-            path.starts_with(locald_root()),
-            "the record has to live in the state root, not beside the app: {}",
-            path.display()
-        );
-        assert_eq!(
-            path.file_name().and_then(std::ffi::OsStr::to_str),
-            Some("shell-update.json"),
-        );
-    }
-
-    #[test]
-    fn the_tray_lock_is_not_held_across_main_thread_round_trips() {
-        // Every `set_*` on a menu item blocks until the main thread is free, and
-        // this runs on the locald reader thread -- so holding the lock across
-        // them stopped daemon events being read whenever the main thread was
-        // busy. Progress froze and `ready` was never handled.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let body = function_body(&source, "fn refresh_agent_host_tray(");
-        let guard_end = body
-            .find("guard.clone()")
-            .expect("the handles are cloned out of the guard");
-        let first_set = body.find(".set_text(").expect("the tray text is set");
-        assert!(
-            guard_end < first_set,
-            "clone the menu handles and drop the guard before any set_* call"
-        );
-    }
-
-    #[test]
-    fn every_command_that_is_not_pure_ui_is_async() {
-        // A synchronous `#[tauri::command]` is dispatched on the main thread, so
-        // one that waits on the daemon, the network or a child process freezes
-        // every window for its whole duration. That is what made a first launch
-        // a black, unresponsive app for minutes.
-        //
-        // Derived from the handler list rather than a list of names, because the
-        // version of this test that named eight commands could not see the
-        // twenty that were still blocking. Adding a synchronous command is now a
-        // decision someone has to write down here, not an omission.
-        //
-        // The allowlist is only for commands that do nothing but touch UI, which
-        // must stay on the main thread.
-        const PURE_UI: &[&str] = &[
-            "open_developer_tools",
-            "close_local_settings",
-            "get_state",
-            // One mutex read of state locald has already pushed into the
-            // shell. It talks to nothing, so dispatching it on the main
-            // thread costs the lock and nothing else -- and it is polled
-            // while a download runs, which is exactly when a command that
-            // waited on the daemon would be felt.
-            "sandbox_image_status",
-        ];
-
-        // Normalised, because the Windows runner checks out CRLF and the
-        // patterns below are written with \n -- which is how this test passed on
-        // macOS and failed on Windows against identical source.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let source = source.as_str();
-        let handlers = {
-            let start = source
-                .find("tauri::generate_handler![")
-                .expect("the handler list exists");
-            let end = source[start..].find("])").expect("the handler list closes") + start;
-            &source[start..end]
-        };
-        let registered: Vec<&str> = handlers
-            .lines()
-            .skip(1)
-            .map(|line| line.trim().trim_end_matches(','))
-            .filter(|name| !name.is_empty() && !name.starts_with("//"))
-            .collect();
-        assert!(
-            registered.len() > 20,
-            "parsed only {} commands, so this test is not reading the handler list",
-            registered.len()
-        );
-
-        for name in registered {
-            if PURE_UI.contains(&name) {
-                assert!(
-                    source.contains(&format!("\nfn {name}(")),
-                    "{name} is on the pure-UI allowlist but is async; either it \
-                     waits on something and should come off the list, or the list \
-                     is stale"
-                );
-                continue;
-            }
-            assert!(
-                source.contains(&format!("async fn {name}("))
-                    || source.contains(&format!("#[tauri::command(async)]\nfn {name}(")),
-                "{name} is dispatched on the main thread. Either make it an async \
-                 command that hands its work to spawn_blocking, or add it to \
-                 PURE_UI with a reason."
-            );
-        }
-    }
-
-    #[test]
-    fn a_resume_that_did_not_pan_out_stops_claiming_the_workspace_is_ready() {
-        // The optimistic resume seeds ready/running/url before it has checked
-        // anything, so the workspace can be on screen with no latency. The
-        // splash opens `ui.url` whenever it loads and finds ready set with no
-        // error -- so every path that gives up and shows the splash has to
-        // clear `ready` first, or the two bounce the user between a splash and
-        // a dead workspace.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let setup = {
-            let start = source.find(".setup(move |app| {").expect("setup exists");
-            let end = source[start..]
-                .find("\n        .build(")
-                .or_else(|| source[start..].find("\n        .run("))
-                .map_or(source.len(), |offset| start + offset);
-            &source[start..end]
-        };
-        let resume = {
-            let start = setup
-                .find("if let Some(target) = resume.clone()")
-                .expect("the resume branch exists");
-            let end = setup[start..]
-                .find("\n                } else {")
-                .map_or(setup.len(), |offset| start + offset);
-            &setup[start..end]
-        };
-        assert!(
-            resume.contains("ui.ready = false;"),
-            "giving up on a resume must clear the optimistic ready flag"
-        );
-        assert_eq!(
-            resume.matches("show_splash(&handle);").count(),
-            1,
-            "the resume branch must reach the splash only through stand_down, \
-             which is what clears the state the splash reads"
-        );
-    }
-
-    #[test]
-    fn nothing_in_setup_installs_the_runtime_on_the_main_thread() {
-        // `setup` runs before the event loop begins pumping, so anything slow
-        // there freezes a window that is already on screen. `ensure_locald`
-        // unpacks the runtime on a first run or an upgrade -- hundreds of
-        // megabytes -- and then waits for the daemon, which is how the splash
-        // came to sit on "Starting Lemma." with no progress for minutes.
-        //
-        // Both launch paths, resume and cold start, must hand that to a worker.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let setup = {
-            let start = source.find(".setup(move |app| {").expect("setup exists");
-            let end = source[start..]
-                .find("\n        .build(")
-                .or_else(|| source[start..].find("\n        .run("))
-                .map_or(source.len(), |offset| start + offset);
-            &source[start..end]
-        };
-        for (index, _) in setup.match_indices("ensure_locald(&handle)") {
-            let preceding = &setup[..index];
-            let spawned = preceding.rfind("std::thread::spawn");
-            let closed = preceding.rfind("});");
-            assert!(
-                spawned.is_some() && spawned > closed,
-                "an ensure_locald in setup is not inside a spawned thread"
-            );
-        }
-    }
-
-    #[test]
-    fn local_settings_says_which_integrations_are_set_up() {
-        // Every row's badge read "Optional" whether or not a credential had been
-        // saved, and the only signal that one had been was the placeholder
-        // inside the input -- grey, and invisible until the drawer was opened.
-        // Somebody who had just saved a Deepgram key had no way to see it land.
-        let markup = include_str!("../ui/control.html").replace("\r\n", "\n");
-        let script = include_str!("../ui/control.js").replace("\r\n", "\n");
-        let style = include_str!("../ui/control.css").replace("\r\n", "\n");
-
-        assert!(
-            !markup.contains(">Optional<"),
-            "a badge that says the same word on every row carries nothing"
-        );
-        assert!(
-            markup.contains("data-config-state"),
-            "each row has a slot for its real state"
-        );
-        // Twice: the definition, and a call. Asserting the function merely
-        // exists passes just as happily when nothing invokes it, which is how a
-        // helper ships dead.
-        assert!(
-            script.matches("paintConfigStates(presence)").count() >= 2,
-            "the painter is defined but never called from the fill pass"
-        );
-        // Read off the row's own fields, so a row added to the markup is
-        // described without anyone remembering a table in the script.
-        assert!(
-            script.contains("input[data-secret]"),
-            "presence of a saved secret is part of being configured"
-        );
-        assert!(
-            style.contains(r#"[data-config-state="configured"]"#),
-            "a configured row has to look different, not just read differently"
-        );
-        // A different axis, and it survives: these rows need a reachable URL
-        // whether or not anyone has filled them in.
-        assert!(
-            markup.contains("Public link"),
-            "the ingress requirement is not a state and should not be replaced by one"
-        );
-    }
-
-    #[test]
-    fn local_settings_never_gates_a_button_on_a_webview_confirm() {
-        // WKWebView routes window.confirm() through a WKUIDelegate panel wry
-        // does not implement, so it returns false without drawing anything:
-        // the click is received and discarded, and the button looks inert.
-        // Destructive actions go through the native dialog command instead.
-        let script = include_str!("../ui/control.js").replace("\r\n", "\n");
-        assert!(
-            !script.contains("window.confirm("),
-            "a destructive button is gated on a confirm() that always says no"
-        );
-        assert!(
-            script.contains("confirm_destructive_action"),
-            "the native confirmation command is how those buttons ask"
-        );
-    }
-
-    /// The source of one free function, for the cases where asserting on
-    /// behaviour would need a running AppHandle.
-    fn function_body<'a>(source: &'a str, signature: &str) -> &'a str {
-        let start = source.find(signature).expect("the function exists");
-        let end = source[start..]
-            .find("\nfn ")
-            .map_or(source.len(), |offset| start + offset);
-        &source[start..end]
-    }
-
-    /// The first screen a user sees is the product's colour, in both themes.
-    ///
-    /// `control.css` rebound its accent to violet and the splash was left
-    /// behind, so the very first impression was a gold-and-cream screen and the
-    /// moment the workspace opened the product was violet. The splash also had
-    /// no dark palette at all, so on a dark-mode Mac -- most of them -- every
-    /// launch, every error and every shutdown flashed a full screen of cream.
-    #[test]
-    fn the_splash_is_the_products_colour_and_follows_the_system_theme() {
-        let splash = include_str!("../ui/index.html").replace("\r\n", "\n");
-
-        for gold in [
-            "#c0801f",
-            "#8a5c16",
-            "0xd89b3d",
-            "216, 155, 61",
-            "192, 128, 31",
-        ] {
-            assert!(
-                !splash.contains(gold),
-                "{gold} is the old accent; the product is violet",
-            );
-        }
-        assert!(splash.contains("@media (prefers-color-scheme: dark)"));
-        assert!(
-            splash.contains("color-scheme: light dark"),
-            "form controls and scrollbars follow the theme too",
-        );
-        // The classic half-themed bug: a colour whose only definition sits
-        // inside the dark block is absent in light mode. Every token the dark
-        // block redefines must also exist on the bare `:root`.
-        let dark_block = splash
-            .split("@media (prefers-color-scheme: dark)")
-            .nth(1)
-            .expect("the dark block exists");
-        let dark_block = &dark_block[..dark_block.find("\n  }\n").unwrap_or(dark_block.len())];
-        let root_block = splash
-            .split(":root {")
-            .nth(1)
-            .expect("the light block exists");
-        for line in dark_block.lines() {
-            let Some(name) = line.trim().strip_prefix("--") else {
-                continue;
-            };
-            let Some(name) = name.split(':').next() else {
-                continue;
-            };
-            assert!(
-                root_block.contains(&format!("--{name}:")),
-                "--{name} is defined only in dark mode, so light mode has no value for it",
-            );
-        }
-    }
-
-    /// Each channel must read the feed that actually carries its builds.
-    ///
-    /// GitHub resolves `releases/latest` to the newest *non-prerelease*.
-    /// Nightlies are prereleases, so a nightly pointed there sees either
-    /// nothing or a stable build -- and an update path that only ever runs on
-    /// release day is one nobody has tested. The nightly tag is rewritten in
-    /// place so the address stays put while its contents move; that fixed
-    /// address is the whole reason a nightly feed can be called durable.
-    #[test]
-    fn each_channel_reads_its_own_feed() {
-        let nightly = updater_endpoints("nightly");
-        let stable = updater_endpoints("stable");
-
-        assert_eq!(nightly.len(), 1);
-        assert!(
-            nightly[0].ends_with("/releases/download/desktop-nightly/latest.json"),
-            "a nightly must read a tag that does not move: {nightly:?}",
-        );
-        assert!(
-            !nightly[0].contains("/releases/latest/"),
-            "`releases/latest` skips prereleases, so it never lists a nightly",
-        );
-
-        assert_eq!(stable, updater_endpoints("dev"), "dev falls back to stable");
-        assert!(stable[0].ends_with("/releases/latest/download/latest.json"));
-
-        // Whatever they are, the plugin has to be able to parse them, and a
-        // typo here would otherwise be a silently empty endpoint list.
-        for channel in ["stable", "nightly", "dev"] {
-            for endpoint in updater_endpoints(channel) {
-                assert!(
-                    tauri::Url::parse(&endpoint).is_ok(),
-                    "{channel} endpoint is not a URL: {endpoint}",
-                );
-            }
-        }
-    }
-
-    /// developer's `cargo tauri dev` -- must never follow an update feed, and
-    /// must say so rather than appearing configured.
-    #[test]
-    fn a_released_build_updates_itself_and_an_unstamped_one_does_not() {
-        // Both halves of this used to be true by construction: tests are a
-        // debug build so `updates_enabled()` is false whatever the channel is,
-        // and `matches!` against the same closed set the `match` produces
-        // cannot fail. It passed with the feature deleted.
-        //
-        // Asserted on the mapping instead, which is the decision worth pinning:
-        // an unrecognised stamp is `dev`, not "assume the best".
-        assert_eq!(channel_of(Some("stable")), "stable");
-        assert_eq!(channel_of(Some("nightly")), "nightly");
-        assert_eq!(channel_of(None), "dev", "an unstamped build is not stable");
-        for unknown in ["Stable", "STABLE", "beta", "", "stable "] {
-            assert_eq!(
-                channel_of(Some(unknown)),
-                "dev",
-                "{unknown:?} must not be read as a release channel",
-            );
-        }
-
-        // And the gate is a conjunction, so nightly cannot update itself even
-        // in a release build.
-        assert!(updates_allowed("nightly", false, true));
-        assert!(!updates_allowed("dev", false, true));
-        assert!(
-            !updates_allowed("stable", true, true),
-            "a debug build never does"
-        );
-        assert!(
-            !updates_allowed("stable", false, false),
-            "and neither does one with no key to verify with",
-        );
-        assert!(updates_allowed("stable", false, true));
-    }
-
-    /// The feed's compatibility block decides before anything is downloaded.
-    #[test]
-    fn an_update_that_would_strand_local_data_says_so_first() {
-        let same = LemmaUpdateMetadata {
-            postgres_major: Some(18),
-            runtime_download_bytes: Some(531_000_000),
-        };
-        assert_eq!(same.compatibility_with(Some(18)), "compatible");
-        assert_eq!(
-            same.compatibility_with(Some(16)),
-            "migration-unavailable",
-            "a Postgres major bump cannot open the existing cluster"
-        );
-
-        // The caller blocks an unknown pairing if an installed runtime has data.
-        assert_eq!(same.compatibility_with(None), "unknown");
-        assert_eq!(
-            LemmaUpdateMetadata::default().compatibility_with(Some(18)),
-            "unknown"
-        );
-    }
-
-    #[test]
-    fn update_preflight_rejects_data_reset_and_unsupported_migrations() {
-        for windows in [false, true] {
-            for has_runtime in [false, true] {
-                assert!(
-                    ensure_update_preserves_data(true, has_runtime, "compatible", windows).is_err()
-                );
-            }
-        }
-        for compatibility in ["unknown", "requires-reset", "migration-unavailable"] {
-            assert!(ensure_update_preserves_data(false, true, compatibility, false).is_err());
-        }
-        assert!(ensure_update_preserves_data(false, true, "compatible", true).is_err());
-        assert!(ensure_update_preserves_data(false, true, "compatible", false).is_ok());
-        assert!(ensure_update_preserves_data(false, false, "unknown", true).is_ok());
-    }
-
-    /// A feed without the block, or with junk in it, is read safely.
-    #[test]
-    fn update_metadata_tolerates_a_feed_that_does_not_carry_it() {
-        let absent = lemma_update_metadata(&json!({"version": "0.8.0"}));
-        assert_eq!(absent.postgres_major, None);
-        assert_eq!(absent.runtime_download_bytes, None);
-
-        let wrong_types = lemma_update_metadata(&json!({
-            "lemma": {"postgres_major": "eighteen", "runtime_download_bytes": []}
-        }));
-        assert_eq!(wrong_types.postgres_major, None);
-        assert_eq!(wrong_types.runtime_download_bytes, None);
-
-        let good = lemma_update_metadata(&json!({
-            "lemma": {"postgres_major": 18, "runtime_download_bytes": 531_000_000_u64}
-        }));
-        assert_eq!(good.postgres_major, Some(18));
-        assert_eq!(good.runtime_download_bytes, Some(531_000_000));
-    }
-
-    #[test]
-    fn update_metadata_uses_the_selected_platform_and_never_another_platforms_fallback() {
-        let feed = json!({"lemma": {
-            "postgres_major": 16, "runtime_download_bytes": 10,
-            "platforms": {
-                "darwin-aarch64": {"postgres_major": 18, "runtime_download_bytes": 20},
-                "windows-x86_64": {"postgres_major": 18, "runtime_download_bytes": 30}
-            }
-        }});
-        assert_eq!(
-            lemma_update_metadata_for(&feed, "darwin-aarch64").runtime_download_bytes,
-            Some(20)
-        );
-        let windows = lemma_update_metadata_for(&feed, "windows-x86_64");
-        assert_eq!(windows.runtime_download_bytes, Some(30));
-        assert_eq!(windows.postgres_major, Some(18));
-        assert_eq!(
-            lemma_update_metadata_for(&feed, "missing-target").postgres_major,
-            None
-        );
-    }
-
-    /// The updater is never reachable from a remote origin.
-    ///
-    /// `workspace.json` grants commands to the locald-served app URL and to
-    /// lemma.work. A permission that can replace the application binary,
-    /// reachable from a page served over the network, would be a
-    /// remote-code-execution primitive -- so the grant lives only on the
-    /// control window, and nothing may hand the plugin's own permissions to
-    /// anyone.
-    #[test]
-    fn no_capability_exposes_the_updater_to_a_remote_origin() {
-        for (name, source) in [
-            ("main", include_str!("../capabilities/main.json")),
-            ("control", include_str!("../capabilities/control.json")),
-            ("workspace", include_str!("../capabilities/workspace.json")),
-        ] {
-            assert!(
-                !source.contains("\"updater:"),
-                "{name} must not grant the updater plugin's own permissions",
-            );
-            assert!(
-                !source.contains("\"process:"),
-                "{name} must not grant process control",
-            );
-        }
-        let workspace = include_str!("../capabilities/workspace.json").replace("\r\n", "\n");
-        for command in ["allow-check-for-app-update", "allow-install-app-update"] {
-            assert!(
-                !workspace.contains(command),
-                "a remote origin must not be able to replace the application",
-            );
-        }
-        assert!(include_str!("../capabilities/control.json").contains("allow-install-app-update"));
-    }
-
-    /// The updater's transport policy is not weakened, and the artifact flag
-    /// stays out of the base config.
-    /// The key an installed app verifies with is real, and matches its own id.
-    ///
-    /// `pubkey: ""` is not a permissive setting. `verify_signature` decodes it
-    /// and fails, so a build carrying an empty key offers an update, downloads
-    /// it, and then cannot verify a thing -- and `updates_enabled()` exists to
-    /// keep such a build from offering one at all.
-    ///
-    /// What was missing is the assertion that the key is *there*. The release
-    /// workflow required the private half and never looked at the public one,
-    /// and the config test asserted the endpoint and the CSP and not this. So a
-    /// release could ship with no key, which is the state this repo was in.
-    #[test]
-    fn the_shipped_public_key_is_a_real_minisign_key() {
-        use base64::Engine as _;
-
-        let config: Value =
-            serde_json::from_str(include_str!("../tauri.conf.json")).expect("config parses");
-        let pubkey = config
-            .pointer("/plugins/updater/pubkey")
-            .and_then(Value::as_str)
-            .expect("the updater declares a pubkey field");
-
-        assert!(
-            !pubkey.trim().is_empty(),
-            "no public key: every install would download an update and then refuse it",
-        );
-        assert!(
-            updater_key_configured(),
-            "the gate that reads this must agree with it",
-        );
-
-        let decoded = base64::engine::general_purpose::STANDARD
-            .decode(pubkey)
-            .expect("the pubkey is base64");
-        let text = String::from_utf8(decoded).expect("a minisign key is text");
-        let mut lines = text.lines();
-        let comment = lines.next().unwrap_or_default();
-        assert!(
-            comment.starts_with("untrusted comment:") && comment.contains("public key"),
-            "this is not a minisign public key: {comment}",
-        );
-
-        // The body carries the algorithm and the key id the signature must
-        // name. A truncated paste passes every check above and none of these.
-        let body = base64::engine::general_purpose::STANDARD
-            .decode(lines.next().expect("a key line follows the comment"))
-            .expect("the key line is base64");
-        assert_eq!(&body[..2], b"Ed", "only Ed25519 keys are signed against");
-        assert_eq!(body.len(), 42, "a minisign public key is 2 + 8 + 32 bytes");
-        let key_id = &body[2..10];
-        assert!(
-            key_id.iter().any(|byte| *byte != 0),
-            "a zero key id is a placeholder, not a key",
-        );
-
-        // And it is not a secret key committed by mistake, which is the same
-        // file format with a different comment.
-        assert!(
-            !text.contains("secret key"),
-            "the SECRET key is in this config; rotate it immediately",
-        );
-    }
-
-    #[test]
-    fn the_updater_config_keeps_its_transport_and_build_boundaries() {
-        let config = include_str!("../tauri.conf.json").replace("\r\n", "\n");
-        assert!(config.contains("\"updater\""));
-        assert!(
-            !config.contains("dangerousInsecureTransportProtocol"),
-            "an update served over plain HTTP is an update anyone can forge",
-        );
-        assert!(
-            config.contains("https://github.com/"),
-            "every endpoint must be HTTPS",
-        );
-        // `tauri build` runs from five places with no signing key -- CI's build
-        // check, the Windows check, two nightly jobs and `make desktop-dmg` --
-        // and every one of them fails if the base config demands updater
-        // artifacts. The flag belongs only in the overlay the release merges.
-        assert!(
-            !config.contains("createUpdaterArtifacts"),
-            "the artifact flag belongs in tauri.updater.conf.json, not the base config",
-        );
-        assert!(include_str!("../tauri.updater.conf.json").contains("createUpdaterArtifacts"));
-
-        // The CSP is untouched: the check runs in Rust precisely so the webview
-        // that also hosts the remote workspace never gains github.com.
-        assert!(
-            config.contains("connect-src 'self' ipc: http://ipc.localhost"),
-            "checking from Rust is what keeps the webview's network policy narrow",
-        );
-    }
-
-    /// An update is only offered by a build that could actually install one.
-    ///
-    /// `tauri.conf.json` ships `"pubkey": ""` until the signing keypair exists,
-    /// and an empty key is not a permissive setting: `verify_signature` decodes
-    /// it and errors, so every install fails -- at the last step, after the app
-    /// has stopped the user's daemon to make room. The config test above passed
-    /// throughout, because it never looked at the key.
-    ///
-    /// Two independent guards, because they fail at opposite ends: this one
-    /// stops the *app* offering what it cannot verify, and the release workflow
-    /// refuses to build at all with the key empty.
-    #[test]
-    fn a_build_with_no_verification_key_does_not_offer_updates() {
-        let configured = updater_key_configured();
-        let committed = serde_json::from_str::<Value>(include_str!("../tauri.conf.json"))
-            .expect("the config parses")
-            .pointer("/plugins/updater/pubkey")
-            .and_then(Value::as_str)
-            .map(|key| !key.trim().is_empty())
-            .expect("the config declares an updater pubkey field");
-        assert_eq!(
-            configured, committed,
-            "the runtime gate must read the key the build actually ships",
-        );
-        if !configured {
-            assert!(
-                !updates_enabled(),
-                "a build that cannot verify an update must not offer one",
-            );
-        }
-    }
-
-    /// The daemon is stopped only once the update is in hand.
-    ///
-    /// `download` is where the signature is verified and where a network drop,
-    /// a moved feed or an undecodable key surfaces. Stopping first meant every
-    /// one of those took the user's whole stack down and then reported a
-    /// failure, for an update that never began.
-    #[test]
-    fn an_update_is_downloaded_and_verified_before_the_stack_is_stopped() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let start = source
-            .find("async fn install_app_update(")
-            .expect("install_app_update exists");
-        let body = &source[start..start + 4200];
-
-        let downloaded = body.find(".download(").expect("it downloads");
-        let stopped = body
-            .find("stop_locald_for_runtime_maintenance")
-            .expect("it stops locald");
-        let installed = body.find(".install(bytes)").expect("it installs");
-        assert!(
-            downloaded < stopped && stopped < installed,
-            "order must be download, stop, install -- got download@{downloaded} \
-             stop@{stopped} install@{installed}",
-        );
-        // A call, not the word -- the comment above the download explains why
-        // the combined form is not used, and would otherwise trip this.
-        assert!(
-            !body.contains(".download_and_install("),
-            "the combined call gives no seam to stop the daemon between the two",
-        );
-    }
-
-    /// Install failures say what happened and what to try.
-    ///
-    /// This mapper had one branch, and it told the reader to "publish its
-    /// runtime artifacts" -- an instruction to a maintainer, shipped to
-    /// strangers. Everything else reached the error screen verbatim.
-    #[test]
-    fn install_failures_are_explained_rather_than_dumped() {
-        let cases = [
-            ("artifact download failed with HTTP 404", "superseded"),
-            (
-                "artifact download failed with HTTP 403",
-                "proxy or firewall",
-            ),
-            ("artifact download failed with HTTP 429", "rate-limited"),
-            (
-                "could not connect to the artifact host",
-                "internet connection",
-            ),
-            ("the request timed out", "resumes from where it stopped"),
-            (
-                "artifact size or SHA-256 did not match the signed manifest",
-                "captive Wi-Fi portal",
-            ),
-            (
-                "signed runtime release 0.9.0 does not match desktop release 0.7.0",
-                "do not match",
-            ),
-        ];
-        for (raw, expected) in cases {
-            let shown = actionable_runtime_install_error(raw);
-            assert!(
-                shown.contains(expected),
-                "{raw:?} should explain {expected:?}, got {shown:?}"
-            );
-            assert!(
-                !shown.contains("HTTP 4"),
-                "a status code is not an explanation: {shown:?}"
-            );
-        }
-
-        // Nobody outside this team can act on either of these.
-        let shown = actionable_runtime_install_error("artifact download failed with HTTP 404");
-        assert!(!shown.contains("Publish"), "{shown:?}");
-        assert!(!shown.contains("PR test DMG"), "{shown:?}");
-
-        // A message that is already specific keeps its numbers.
-        let disk =
-            "not enough disk space for Lemma's local runtime: 7 GiB required, 3 GiB available";
-        assert_eq!(actionable_runtime_install_error(disk), disk);
-    }
-
-    /// Diagnostics masks the secrets it has never seen.
-    ///
-    /// Every operator secret lives in the OS credential vault, and the file
-    /// substitution this used to be could only mask values it had read off
-    /// disk -- so the AI provider key, the Slack and Telegram tokens and the
-    /// OAuth client secrets were unredactable by construction, while the
-    /// README promised bounded, redacted output.
-    #[test]
-    fn diagnostics_masks_credentials_it_has_never_seen() {
-        let masked = mask_secret_shapes(
-            [
-                "provider rejected key sk-EXAMPLE-NOT-A-REAL-KEY-FOR-TESTS",
-                "slack bot xoxb-EXAMPLE-NOT-A-REAL-SLACK-TOKEN responded 200",
-                "GET /v1/models Authorization: Bearer eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxIn0.c2ln",
-                "session eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhYmMifQ.QWxhZGRpbjpvcGVu expired",
-                "resend re_EXAMPLE-NOT-A-REAL-RESEND-KEY accepted",
-            ]
-            .join("\n"),
-        );
-
-        for leaked in [
-            "sk-EXAMPLE-NOT-A-REAL-KEY-FOR-TESTS",
-            "xoxb-EXAMPLE-NOT-A-REAL-SLACK-TOKEN",
-            "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhYmMifQ.QWxhZGRpbjpvcGVu",
-            "re_EXAMPLE-NOT-A-REAL-RESEND-KEY",
-        ] {
-            assert!(!masked.contains(leaked), "{leaked} survived redaction");
-        }
-        assert!(
-            !masked.contains("Bearer eyJ"),
-            "an Authorization header carries the credential in full: {masked}"
-        );
-        // Still a diagnostic afterwards. A log masked so heavily that nobody
-        // can read it does not help the person reading it.
-        assert!(masked.contains("provider rejected key"));
-        assert!(masked.contains("responded 200"));
-        assert!(masked.contains("expired"));
-    }
-
-    /// A source-searching test must not depend on how the repo was checked out.
-    ///
-    /// Git's default on Windows rewrites text files to CRLF, and these tests
-    /// read their own source and the bundled UI through `include_str!`. A
-    /// needle containing `\n` then matches nothing -- but only sometimes:
-    /// `find("\nfn ")` still matches inside `"\r\nfn "`, so most survived and
-    /// exactly two did not. The failures appear only on the Windows job, which
-    /// is not in the desktop path filter, so each one costs a push to see.
-    ///
-    /// Two defences, and this asserts the one that can be asserted from here:
-    /// every `include_str!` bound for searching normalises on the way in.
-    /// `.gitattributes` is the other, and means a checkout never has CRLF to
-    /// normalise.
-    #[test]
-    fn every_included_source_is_read_with_normalised_line_endings() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let mut unnormalised = Vec::new();
-        for (number, line) in source.lines().enumerate() {
-            let trimmed = line.trim();
-            // A binding, which is what gets searched. An inline
-            // `include_str!(..).contains("one line")` cannot span a newline.
-            if !trimmed.starts_with("let ") || !trimmed.contains("= include_str!(") {
-                continue;
-            }
-            if !trimmed.contains(r#".replace("\r\n", "\n")"#) {
-                unnormalised.push(format!("main.rs:{}: {trimmed}", number + 1));
-            }
-        }
-        assert!(
-            unnormalised.is_empty(),
-            "these read included text without normalising, so a needle \
-             containing a newline finds nothing on a Windows checkout:\n{}",
-            unnormalised.join("\n"),
-        );
-    }
-
-    /// Redaction does not reformat the log on the way past.
-    ///
-    /// The masker used to rebuild each line with
-    /// `split_whitespace().join(" ")`, which redacts correctly and flattens
-    /// everything else: indentation, tabs, aligned columns. Diagnostics is
-    /// where somebody reads a Python traceback, and a traceback with no
-    /// indentation is a wall of text.
-    ///
-    /// The neighbouring test could not see this -- every line in its fixture is
-    /// single-spaced, so collapsing runs of whitespace is a no-op on it.
-    #[test]
-    fn redaction_leaves_the_shape_of_a_traceback_alone() {
-        let traceback = concat!(
-            "Traceback (most recent call last):\n",
-            "  File \"/app/main.py\", line 42, in handler\n",
-            "    raise RuntimeError(\"boom\")\n",
-            "\tRuntimeError: boom\n",
-            "  key   sk-ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789abcdef\n",
-        );
-
-        let masked = mask_secret_shapes(traceback.to_owned());
-
-        assert!(
-            masked.contains("  File \"/app/main.py\", line 42, in handler"),
-            "two-space indentation is what makes a traceback readable:\n{masked}",
-        );
-        assert!(
-            masked.contains("    raise RuntimeError"),
-            "four-space indentation is gone:\n{masked}",
-        );
-        assert!(masked.contains('\t'), "a tab is whitespace too:\n{masked}");
-        assert!(
-            masked.contains("  key   [redacted]"),
-            "the run of spaces between a label and its value is alignment:\n{masked}",
-        );
-        // And the point of the exercise still holds.
-        assert!(!masked.contains("sk-ABCDEFGHIJ"), "{masked}");
-    }
-
-    /// Redaction must not eat the things a log is read for.
-    #[test]
-    fn diagnostics_keeps_paths_versions_and_digests_readable() {
-        let text = [
-            "installed /Applications/Lemma.app/Contents/MacOS/lemma-locald",
-            "release 0.7.0 pinned docker.io/pgvector/pgvector:0.8.3-pg18",
-            "sha256:c8a919765f2ef63681329fa21021b830cd4d79d1165bdca730dd016014e4da84",
-            "listening on http://app.lemma.localhost:49180",
-        ]
-        .join("\n");
-
-        assert_eq!(
-            mask_secret_shapes(text.clone()),
-            text,
-            "paths, versions and digests are not credentials"
-        );
-    }
-
-    /// A release build honours no runtime-redirecting environment variable.
-    ///
-    /// These three point the app at a different host pack, managed runtime, or
-    /// signed manifest -- and the manifest carries the digests everything else
-    /// is verified against, so overriding it chooses both the bytes and the
-    /// check on them. Read unconditionally, they let anything already running
-    /// as the user make a notarized, hardened-runtime Lemma fetch and execute a
-    /// runtime of its choosing.
-    ///
-    /// Asserted on the source because the property is "the read is gated", and
-    /// a test running under `cfg(test)` is a debug build -- it cannot observe
-    /// the release behaviour by calling the function.
-    #[test]
-    fn a_release_build_ignores_every_runtime_redirecting_env_var() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        for name in [
-            "LEMMA_DESKTOP_HOST_PACK_ROOT",
-            "LEMMA_DESKTOP_MANAGED_RUNTIME_ROOT",
-            "LEMMA_DESKTOP_RELEASE_MANIFEST",
-        ] {
-            assert!(
-                !source.contains(&format!("std::env::var_os(\"{name}\")")),
-                "{name} must be read through dev_override, which is inert in a \
-                 release build, not directly",
-            );
-            assert!(
-                source.contains(&format!("dev_override(\"{name}\")")),
-                "{name} should still be honoured in a development build",
-            );
-        }
-        // The gate itself, so this cannot pass against a `dev_override` that
-        // forgot to check.
-        assert!(
-            function_body(&source, "fn dev_override(name: &str)")
-                .contains("if !cfg!(debug_assertions)"),
-            "dev_override must be inert outside a development build",
-        );
-    }
-
-    #[test]
-    fn a_busy_daemon_refuses_an_operation_rather_than_dropping_it() {
-        let mut ui = UiState::default();
-        reserve_ui_operation(&mut ui, "start", "starting").unwrap();
-        assert_eq!(
-            reserve_ui_operation(&mut ui, "restart", "retry"),
-            Err(LOCALD_BUSY.into())
-        );
-        assert_eq!(ui.active_operation_id, "starting");
-        reserve_ui_operation(&mut ui, "shutdown-daemon", "quitting").unwrap();
-        assert_eq!(ui.active_operation_id, "quitting");
-        assert!(ui.completed_operation_ids.iter().any(|id| id == "starting"));
-        assert_eq!(
-            reserve_ui_operation(&mut ui, "start", "late-start"),
-            Err(LOCALD_BUSY.into())
-        );
-    }
-
-    #[test]
-    fn shutdown_keeps_its_progress_when_a_background_health_probe_fails() {
-        assert!(!event_applies_during_shutdown(&json!({
-            "event": "error", "code": "managed-runtime-lost", "message": "health deadline",
-        })));
-        assert!(!event_applies_during_shutdown(&json!({
-            "event": "ready", "operation_id": "old-start",
-        })));
-        assert!(event_applies_during_shutdown(&json!({
-            "event": "phase", "operation_id": "quit", "label": "Stopping Lemma",
-        })));
-        assert!(event_applies_during_shutdown(&json!({
-            "event": "error", "id": "quit", "code": "shutdown-failed",
-        })));
-        assert!(event_applies_during_shutdown(&json!({
-            "event": "done", "id": "quit", "cmd": "shutdown-daemon", "ok": true,
-        })));
-    }
-
-    #[test]
-    fn a_refused_stop_leaves_no_splash_behind() {
-        // The splash used to go up before the stop was sent, so a refusal left
-        // a "stopping Lemma" screen in front of a stack nobody had asked to
-        // stop, with no way back.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let body = function_body(&source, "fn stop_impl(");
-        let sent = body.find("send_local_operation").expect("stop_impl sends");
-        let splash = body
-            .find("show_splash_with_intent")
-            .expect("stop_impl shows a splash");
-        assert!(
-            sent < splash,
-            "the splash must follow the accepted operation, not precede it"
-        );
-    }
-
-    #[test]
-    fn giving_up_on_quit_says_so() {
-        // Confirming "Stop and Quit" and then getting neither, silently, is the
-        // failure this guards.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let body = function_body(&source, "fn stop_then_quit(");
-        assert!(
-            body.contains("report_action_failure"),
-            "a quit that cannot start its stop must tell the user why"
-        );
-    }
-
-    #[test]
-    fn a_growing_log_keeps_its_identity_so_the_tail_cursor_survives() {
-        // The identity exists to answer "is this still the same file", and the
-        // cursor is thrown away whenever it changes. Deriving it from anything
-        // that grows with the log -- size, last write time -- compiles fine and
-        // then re-sends the whole tail on every poll of an active log, i.e. it
-        // breaks precisely when someone is watching a failure happen.
-        use std::io::Write;
-
-        let directory = tempfile::tempdir().expect("temp dir");
-        let path = directory.path().join("backend.log");
-        std::fs::write(&path, b"first\n").expect("seed the log");
-
-        let before = super::diagnostic_file_identity(&File::open(&path).expect("open"));
-        assert_ne!(
-            before, "",
-            "the identity should be readable on this platform"
-        );
-
-        let mut appended = std::fs::OpenOptions::new()
-            .append(true)
-            .open(&path)
-            .expect("append to the log");
-        appended.write_all(b"second\n").expect("write");
-        appended.flush().expect("flush");
-        drop(appended);
-
-        let after = super::diagnostic_file_identity(&File::open(&path).expect("reopen"));
-        assert_eq!(before, after, "appending to a log must not re-identify it");
-    }
-
-    #[test]
-    fn replacing_a_log_changes_its_identity_so_a_stale_cursor_is_dropped() {
-        let directory = tempfile::tempdir().expect("temp dir");
-        let path = directory.path().join("backend.log");
-        std::fs::write(&path, b"old\n").expect("seed the log");
-        let rotated = super::diagnostic_file_identity(&File::open(&path).expect("open"));
-
-        std::fs::rename(&path, directory.path().join("backend.log.1")).expect("rotate");
-        std::fs::write(&path, b"new\n").expect("fresh log");
-
-        let fresh = super::diagnostic_file_identity(&File::open(&path).expect("reopen"));
-        assert_ne!(
-            rotated, fresh,
-            "a rotated log is a different file and must reset the cursor"
-        );
-    }
-    use super::*;
-
-    fn capability(name: &str) -> Value {
-        let raw = match name {
-            "main" => include_str!("../capabilities/main.json"),
-            "control" => include_str!("../capabilities/control.json"),
-            "confirmation" => include_str!("../capabilities/confirmation.json"),
-            "workspace" => include_str!("../capabilities/workspace.json"),
-            other => panic!("unknown capability {other}"),
-        };
-        serde_json::from_str(raw).expect("capability is valid JSON")
-    }
-
-    fn granted(name: &str) -> Vec<String> {
-        capability(name)["permissions"]
-            .as_array()
-            .expect("permissions array")
-            .iter()
-            .filter_map(|value| value.as_str().map(str::to_string))
-            .collect()
-    }
-
-    #[test]
-    fn every_command_is_granted_to_exactly_the_surfaces_that_call_it() {
-        // Declaring an app manifest means an ungranted command is rejected at
-        // runtime, from the bundled pages too. Nothing in the build surfaces
-        // that - the page just stops working - so pin it here instead.
-        let commands = include_str!("../build.rs").replace("\r\n", "\n");
-        let all: Vec<String> = commands
-            .lines()
-            .filter_map(|line| {
-                line.trim()
-                    .strip_prefix('"')?
-                    .strip_suffix("\",")
-                    .map(str::to_string)
-            })
-            .collect();
-        assert!(
-            all.len() > 15,
-            "failed to parse the command list from build.rs"
-        );
-
-        let mut all_granted: Vec<String> = Vec::new();
-        for name in ["main", "control", "workspace", "confirmation"] {
-            all_granted.extend(granted(name));
-        }
-        for command in &all {
-            let permission = format!("allow-{}", command.replace('_', "-"));
-            assert!(
-                all_granted.contains(&permission),
-                "{command} is registered but no capability grants {permission}, so every call to it is rejected",
-            );
-        }
-
-        // Granted *somewhere* is not enough: a capability names one webview, so
-        // a page calling a command only its sibling was granted is still
-        // rejected at runtime. Check each bundled page against its own grants.
-        for (capability, script) in [
-            ("main", include_str!("../ui/index.html")),
-            ("control", include_str!("../ui/control.js")),
-            ("confirmation", include_str!("../ui/confirmation.js")),
-        ] {
-            let grants = granted(capability);
-            for command in invoked_commands(script) {
-                let permission = format!("allow-{}", command.replace('_', "-"));
-                assert!(
-                    grants.contains(&permission),
-                    "{capability} calls {command} but its capability lacks {permission}",
-                );
-            }
-        }
-    }
-
-    /// Every `invoke("name")` a bundled page makes.
-    fn invoked_commands(script: &str) -> Vec<String> {
-        script
-            .split("invoke(\"")
-            .skip(1)
-            .filter_map(|rest| rest.split('"').next().map(str::to_string))
-            .collect()
-    }
-
-    #[test]
-    fn a_daemon_from_a_replaced_app_bundle_is_not_this_build() {
-        let expected = std::path::Path::new("/Applications/Lemma.app/Contents/MacOS/lemma-locald");
-        let this_build = json!({
-            "daemon_api_revision": REQUIRED_LOCALD_API_REVISION,
-            "executable": path_identity(expected),
-        });
-        assert!(locald_is_this_build(&this_build, Some(expected)));
-
-        // The failure this exists for. Updating the app moves the previous
-        // bundle to the Trash; its daemon keeps running and keeps the socket,
-        // reporting the same version and the same revision as the one that
-        // replaced it.
-        let from_the_trash = json!({
-            "daemon_api_revision": REQUIRED_LOCALD_API_REVISION,
-            "daemon_version": env!("CARGO_PKG_VERSION"),
-            "executable": "/Users/someone/.Trash/Lemma 4.23.56 PM.app/Contents/MacOS/lemma-locald",
-        });
-        assert!(
-            !locald_is_this_build(&from_the_trash, Some(expected)),
-            "a daemon running from another bundle must be replaced, not adopted",
-        );
-
-        // Every build older than the field predates the check, so silence is a
-        // mismatch rather than a pass.
-        let before_the_field = json!({ "daemon_api_revision": REQUIRED_LOCALD_API_REVISION });
-        assert!(!locald_is_this_build(&before_the_field, Some(expected)));
-
-        let wrong_revision = json!({
-            "daemon_api_revision": REQUIRED_LOCALD_API_REVISION + 1,
-            "executable": path_identity(expected),
-        });
-        assert!(!locald_is_this_build(&wrong_revision, Some(expected)));
-
-        // A source checkout has no packaged sidecar to be, and its daemon is
-        // whatever the developer just built.
-        assert!(locald_is_this_build(&before_the_field, None));
-    }
-
-    #[test]
-    fn switching_connection_says_what_it_is_about_to_do() {
-        let (title, body, confirm) = connection_switch_prompt("hosted", false);
-        assert_eq!(title, format!("Run Lemma on {THIS_COMPUTER}?"));
-        assert_eq!(confirm, "Start Local");
-        // The ninety-second health gate is the whole reason this prompt exists:
-        // the press used to be followed by silence for minutes.
-        assert!(body.contains("takes a few minutes"), "{body}");
-
-        let (title, body, confirm) = connection_switch_prompt("local", true);
-        assert_eq!(title, "Use the hosted workspace?");
-        assert_eq!(confirm, "Use Hosted");
-        assert!(
-            body.contains(&format!("keeps running on {THIS_COMPUTER}")),
-            "{body}"
-        );
-        // Leaving must never read as destroying: the pods stay.
-        assert!(body.contains("stay where they are"), "{body}");
-    }
-
-    #[test]
-    fn the_tray_reports_reachability_rather_than_liveness() {
-        // Each of these is a live process that cannot take work, and the old
-        // process-only status called them all "running".
-        assert_eq!(
-            agent_host_tray_label(true, true, false, false, false, false),
-            "Agent Host: not paired",
-        );
-        assert_eq!(
-            agent_host_tray_label(true, true, true, false, false, false),
-            "Agent Host: reconnecting…",
-        );
-        assert_eq!(
-            agent_host_tray_label(true, true, true, true, false, false),
-            "Agent Host: connected",
-        );
-        // Paired and not running is a stage, not a setting: the off switch is
-        // gone, so this state is only ever on its way to "connected".
-        assert_eq!(
-            agent_host_tray_label(true, false, true, false, false, false),
-            "Agent Host: starting…"
-        );
-        // A build without the sidecar has no host to report on at all.
-        assert_eq!(
-            agent_host_tray_label(false, false, false, false, false, false),
-            "Agent Host: not installed",
-        );
-        // The state this distinction was added for: a workspace that answered
-        // yesterday and is not there today. "Reconnecting" for a week is a
-        // promise the retry loop cannot keep.
-        assert_eq!(
-            agent_host_tray_label(true, true, true, false, true, false),
-            "Agent Host: workspace unreachable",
-        );
-        // Connected wins over a stale error on another target: one workspace
-        // failing does not stop this computer taking work from the other.
-        assert_eq!(
-            agent_host_tray_label(true, true, true, true, true, false),
-            "Agent Host: connected",
-        );
-    }
-
-    #[test]
-    fn a_sidecar_that_cannot_start_does_not_say_it_is_starting() {
-        // "starting…" replaced "off" because nothing can switch this computer
-        // off any more — but it is a promise about what happens next, and the
-        // supervisor arms a restart backoff on every failed spawn. A sidecar
-        // that cannot start at all said "starting…" for as long as the app was
-        // open, with nothing anywhere to contradict it. Same shape as an adapter
-        // stuck at "Setting up": a failure wearing the clothes of progress.
-        assert_eq!(
-            agent_host_tray_label(true, false, true, false, false, true),
-            "Agent Host: not starting — see log",
-        );
-        // A failure recorded against a host that is up is about its workspaces,
-        // not its startup, and the reachability states already say that better.
-        assert_eq!(
-            agent_host_tray_label(true, true, true, true, false, true),
-            "Agent Host: connected",
-        );
-        assert_eq!(
-            agent_host_tray_label(true, true, true, false, true, true),
-            "Agent Host: workspace unreachable",
-        );
-        // And a build with no sidecar has nothing to fail.
-        assert_eq!(
-            agent_host_tray_label(false, false, false, false, false, true),
-            "Agent Host: not installed",
-        );
-    }
-
-    #[test]
-    fn the_workspace_origin_reaches_local_settings_and_nothing_else() {
-        // The workspace is a remote origin to Tauri - locald serves it over
-        // http, and the hosted build loads lemma.work - so without this
-        // capability its Local settings button is silently rejected by the ACL.
-        //
-        // What it may reach is deliberately short: Local settings, this
-        // computer's Agent Host, the AI provider, and how the sandbox image
-        // download is going. The provider one was added because onboarding
-        // cannot honestly ask "which model?" and then send the user to a
-        // different window for the answer — and it is safe to add precisely
-        // because `configure_ai_provider` reaches `config.set-ai`, which merges
-        // that one section. `allow-apply-operator-config`, which would let the
-        // same page rewrite sharing and surfaces, stays out.
-        //
-        // `allow-sandbox-image-status` is the mildest of the four: it reads two
-        // strings the shell already holds and changes nothing at all. It is
-        // here because the download runs behind a workspace the user is already
-        // in, so the workspace is the only surface that can say it is happening.
-        let workspace = granted("workspace");
-        assert!(workspace.contains(&"allow-open-control-center".to_string()));
-        assert!(workspace.iter().all(|permission| {
-            matches!(
-                permission.as_str(),
-                "allow-open-control-center"
-                    | "allow-discover-provider-models"
-                    | "allow-configure-ai-provider"
-                    | "allow-sandbox-image-status"
-            ) || permission.starts_with("allow-agent-host-")
-        }));
-        for forbidden in [
-            "allow-apply-operator-config",
-            "allow-sharing-action",
-            "allow-prepare-runtime",
-            "allow-repair-runtime",
-            "allow-stop",
-            "core:default",
-        ] {
-            assert!(
-                !workspace.contains(&forbidden.to_string()),
-                "the workspace origin must not be granted {forbidden}",
-            );
-        }
-
-        let patterns: Vec<tauri::utils::acl::RemoteUrlPattern> = capability("workspace")["remote"]
-            ["urls"]
-            .as_array()
-            .expect("remote urls")
-            .iter()
-            .map(|value| {
-                value
-                    .as_str()
-                    .expect("url string")
-                    .parse()
-                    .expect("valid pattern")
-            })
-            .collect();
-        let matches = |raw: &str| {
-            let url = tauri::Url::parse(raw).expect("valid url");
-            patterns.iter().any(|pattern| pattern.test(&url))
-        };
-
-        assert!(matches("http://app.lemma.localhost:3711/pod/abc"));
-        assert!(matches("http://app.lemma.localhost:63844/"));
-        assert!(matches("https://lemma.work/pod/abc"));
-
-        // Sharing publishes the same workspace on a different host. Those
-        // visitors must not be able to drive this Mac's Local settings.
-        assert!(!matches("http://192.168.1.24:3711/"));
-        assert!(!matches("https://team.trycloudflare.com/"));
-        assert!(!matches("https://lemma.work.evil.example/"));
-        assert!(!matches("http://lemma.work/"));
-    }
-
-    #[test]
-    fn local_settings_does_not_inherit_the_splash_commands() {
-        // Local settings is a child webview of the main window, and capability
-        // matching is (window OR webview). A window-scoped splash capability
-        // would therefore hand its commands to Local settings as well.
-        for name in ["main", "control", "workspace", "confirmation"] {
-            assert!(
-                capability(name).get("windows").is_none(),
-                "{name} must scope by webview, not window",
-            );
-        }
-        assert_eq!(capability("main")["webviews"][0], "main");
-        assert_eq!(capability("control")["webviews"][0], "control");
-
-        let splash_only = [
-            "allow-prepare-runtime",
-            "allow-choose-connection-mode",
-            "allow-login",
-        ];
-        for permission in splash_only {
-            assert!(granted("main").contains(&permission.to_string()));
-            assert!(!granted("control").contains(&permission.to_string()));
-        }
-        assert!(granted("control").contains(&"allow-apply-operator-config".to_string()));
-        assert!(!granted("main").contains(&"allow-apply-operator-config".to_string()));
-    }
-
-    #[test]
-    fn an_overridden_workspace_origin_gets_the_same_commands() {
-        let capability_for = |values: &[&str]| {
-            workspace_capability_for(values.iter().map(|value| value.to_string()))
-        };
-
-        assert!(capability_for(&[]).is_none());
-        // A shipped origin is already covered; re-granting it would only widen
-        // the pattern set for no reason.
-        assert!(capability_for(&["https://lemma.work"]).is_none());
-        assert!(capability_for(&["not a url"]).is_none());
-
-        // ...including a local workspace, whose port is not known until it is
-        // allocated. The shipped entry is `http://app.<base>:*`, and the origin
-        // checked against it is concrete, so a plain string comparison never
-        // matched one and quietly minted an override for every local launch.
-        for base in ["lemma.localhost", "127.0.0.1.sslip.io"] {
-            assert!(
-                capability_for(&[&format!("http://app.{base}:52413")]).is_none(),
-                "the shipped capability already covers app.{base} on any port",
-            );
-        }
-        // And the wildcard is the port alone. A host that merely starts the
-        // same way is a different machine, and must still be treated as an
-        // override rather than silently accepted as shipped.
-        assert!(
-            capability_for(&["http://app.lemma.localhost.evil:52413"]).is_some(),
-            "a lookalike host must not read as a shipped origin",
-        );
-        assert!(!shipped_workspace_origin_covers(
-            "http://app.lemma.localhost:*",
-            "http://app.lemma.localhost:52413/admin"
-        ));
-        assert!(!shipped_workspace_origin_covers(
-            "http://app.lemma.localhost:*",
-            "http://app.lemma.localhost"
-        ));
-
-        let raw = capability_for(&["https://staging.lemma.work/", "http://127.0.0.1:3711"])
-            .expect("an overridden origin produces a capability");
-        let capability: Value = serde_json::from_str(&raw).expect("valid capability JSON");
-        assert_eq!(
-            capability["remote"]["urls"],
-            json!(["https://staging.lemma.work", "http://127.0.0.1:3711"]),
-        );
-        // Read from the shipped file rather than restated here: a hardcoded copy
-        // is exactly what drifted, and an assertion that has to be remembered
-        // catches nothing.
-        let shipped: Value =
-            serde_json::from_str(SHIPPED_WORKSPACE_CAPABILITY).expect("valid shipped capability");
-        assert_eq!(
-            capability["permissions"], shipped["permissions"],
-            "an override must reach neither further nor less far than the shipped capability",
-        );
-        // The one this drift actually cost, named so the regression reads as
-        // the symptom it produced.
-        assert!(
-            capability["permissions"]
-                .as_array()
-                .expect("permissions are an array")
-                .contains(&json!("allow-agent-host-status")),
-            "a self-hosted workspace must be able to ask this computer for its status",
-        );
-        assert_eq!(capability["local"], json!(false));
-    }
-
-    /// Every local base this build serves has to be an origin the workspace
-    /// capability covers.
-    ///
-    /// The failure this catches is silent and total. A workspace served on a
-    /// base the capability does not list matches no capability at all, so
-    /// opening Local settings, connecting the Agent Host and configuring a
-    /// provider each answer `not allowed by ACL` -- the whole of onboarding,
-    /// with nothing on screen to say why. Nothing else ties the two files
-    /// together: the base domain is chosen in locald and the origins are
-    /// declared in a Tauri capability, and neither imports the other.
-    #[test]
-    fn every_local_base_this_build_serves_is_a_shipped_workspace_origin() {
-        let shipped = shipped_workspace_origins();
-        for base in TRUSTED_LOCAL_BASES {
-            let origin = format!("http://app.{base}:52413");
-            assert!(
-                shipped
-                    .iter()
-                    .any(|pattern| shipped_workspace_origin_covers(pattern, &origin)),
-                "capabilities/workspace.json does not cover {origin}, so a                  workspace served there reaches no shell command at all",
-            );
-        }
-    }
-
-    #[test]
-    fn local_settings_declares_no_palette_of_its_own() {
-        // Local settings is a separate webview for a security reason — the
-        // privileged commands are granted only to it — but that never justified
-        // a second visual identity, which is what this file had: its own gold
-        // accent, its own paper, its own display face. Colours now live in one
-        // token block at the top and every rule reads from it, so this catches
-        // the next raw hex before it becomes a third design system.
-        let css = include_str!("../ui/control.css").replace("\r\n", "\n");
-        let rules = css
-            .split_once("* { box-sizing: border-box; }")
-            .expect("control.css starts with a token block then its rules")
-            .1;
-
-        let mut stray: Vec<&str> = Vec::new();
-        for (index, _) in rules.match_indices('#') {
-            let literal: String = rules[index..]
-                .chars()
-                .take_while(|character| character.is_ascii_hexdigit() || *character == '#')
-                .collect();
-            // A CSS id selector is not a colour; a colour is 3, 4, 6, or 8
-            // hex digits and nothing else.
-            if matches!(literal.len(), 4 | 5 | 7 | 9) && literal != "#ffffff" {
-                stray.push(&rules[index..index + literal.len()]);
-            }
-        }
-        assert!(
-            stray.is_empty(),
-            "control.css rules must read colours from the token block, found: {stray:?}"
-        );
-
-        // And the tokens themselves must be the product's, not a parallel set.
-        assert!(
-            css.contains("--accent-rgb: 90 63 212"),
-            "light accent is the action violet"
-        );
-        assert!(
-            css.contains("--accent-rgb: 139 122 245"),
-            "dark accent is the action violet"
-        );
-        assert!(
-            css.contains("--canvas: #f2efe7"),
-            "light canvas is the product's paper"
-        );
-        assert!(
-            !css.contains("Bricolage"),
-            "the page no longer carries its own display face"
-        );
-    }
-
-    #[test]
-    fn waiting_returns_as_soon_as_the_label_comes_free() {
-        // The event loop needs a moment after `destroy`, not a fixed delay --
-        // so this must return on the first poll that sees the label gone, and
-        // must not have given up before then.
-        let polls = std::cell::Cell::new(0);
-        let released = wait_until_label_released(
-            || {
-                polls.set(polls.get() + 1);
-                polls.get() < 3
-            },
-            Duration::from_secs(5),
-            Duration::from_millis(1),
-        );
-
-        assert!(released);
-        assert_eq!(polls.get(), 3, "it stopped polling the moment it came free");
-    }
-
-    #[test]
-    fn waiting_gives_up_rather_than_hanging_the_switch() {
-        // A label that never frees must not park a blocking thread forever.
-        // The caller logs and tries anyway; what it must not do is never return.
-        let start = Instant::now();
-        let released =
-            wait_until_label_released(|| true, Duration::from_millis(30), Duration::from_millis(1));
-
-        assert!(!released);
-        assert!(start.elapsed() >= Duration::from_millis(30));
-        assert!(
-            start.elapsed() < Duration::from_secs(5),
-            "the timeout has to actually bound the wait"
-        );
-    }
-
-    #[test]
-    fn a_replaced_window_is_measured_before_it_is_destroyed() {
-        // Read order, not a nicety: `outer_position` and `inner_size` need a
-        // window, and the whole point is that there is about to not be one. The
-        // rebuilt window used to come back at the OS default placement in the
-        // default size, which is a swap the user sees even if nothing flashes.
-        //
-        // Asserted on the source for the same reason as the wait below: reaching
-        // the real path needs a running event loop.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let body = {
-            let start = source
-                .find("fn rebuild_main_window_for_mode(app: &AppHandle, mode: &str)")
-                .expect("rebuild_main_window_for_mode exists");
-            let end = source[start..]
-                .find("\nfn ")
-                .map_or(source.len(), |offset| start + offset);
-            &source[start..end]
-        };
-
-        let measured = body
-            .find("placement_of(&existing)")
-            .expect("it reads where the window was");
-        let destroy = body.find(".destroy()").expect("it destroys the old window");
-
-        assert!(
-            measured < destroy,
-            "the window has to be measured while it still exists"
-        );
-        assert!(
-            body.contains(
-                "build_main_window_at(app, mode, initial.clone(), true, true, placement)"
-            ),
-            "the replacement is told it is one, and where to sit"
-        );
-    }
-
-    /// A saved window position must not be able to hide the app.
-    ///
-    /// The failure is the classic one and it is unrecoverable without a
-    /// terminal: quit with the window on a second display, unplug it, launch.
-    /// The window is real, focused, and nowhere on screen, and the only way
-    /// back is deleting a config file the user does not know exists.
-    /// Closing the window keeps Lemma running; quitting stops it completely.
-    ///
-    /// This is the bargain the product makes, and only half of it was true.
-    /// Closing hid to the tray and left everything up, which is right and is
-    /// what the tray icon is for. Quitting exited the app and left
-    /// `lemma-locald` running -- supervising Postgres, Redis, the backend, the
-    /// Agent Host and a virtual machine -- with no window, no tray icon and
-    /// nothing in the Dock. The only way to see it was `ps` and the only way to
-    /// stop it was `kill`.
-    ///
-    /// A background service is fine. A background service with no interface is
-    /// not one anybody agreed to.
-    /// Every redirect the daemon reads is one a release build takes away.
-    ///
-    /// `dev_override` gates the variables the *app* reads, and stops exactly
-    /// there: the daemon inherits this process's environment and reads
-    /// redirects of its own, so a signed build refused to load a host pack from
-    /// an environment variable and then handed that same variable to the
-    /// process that loads it without asking.
-    ///
-    /// Enumerated from locald's own source rather than from a list somebody
-    /// maintains, because a list somebody maintains is how the first three went
-    /// missing. A new variable there is a compile-time-green, review-invisible
-    /// hole until this test names it.
-    #[test]
-    fn every_daemon_redirect_variable_is_stripped_from_a_release_build() {
-        // The daemon is a directory, and it is read from disk rather than
-        // named file by file. Enumerating its modules here would reintroduce
-        // exactly the maintained list this test exists to replace: a module
-        // added there would be compile-time green and review-invisible, which
-        // is how the first three variables went missing.
-        let daemon = std::path::Path::new(env!("CARGO_MANIFEST_DIR")).join("locald/src/daemon");
-        let mut sources: Vec<String> = std::fs::read_dir(&daemon)
-            .expect("locald's daemon module directory")
-            .filter_map(Result::ok)
-            .map(|entry| entry.path())
-            .filter(|path| path.extension().is_some_and(|kind| kind == "rs"))
-            .map(|path| std::fs::read_to_string(&path).expect("a daemon module"))
-            .collect();
-        assert!(
-            sources.len() > 1,
-            "the daemon is a directory of modules; reading one file means the \
-             scan is looking at a fraction of it"
-        );
-        sources.extend(
-            [
-                include_str!("../locald/src/agent_host.rs"),
-                include_str!("../locald/src/native_host_pack.rs"),
-                include_str!("../locald/src/managed_runtime.rs"),
-                include_str!("../locald/src/paths.rs"),
-                include_str!("../local-runtime/manager/src/lib.rs"),
-            ]
-            .into_iter()
-            .map(str::to_owned),
-        );
-
-        let mut read_by_the_daemon = std::collections::BTreeSet::new();
-        for source in &sources {
-            let mut rest = source.as_str();
-            while let Some(at) = rest.find("LEMMA_") {
-                let tail = &rest[at..];
-                let end = tail
-                    .find(|c: char| !c.is_ascii_uppercase() && c != '_' && !c.is_ascii_digit())
-                    .unwrap_or(tail.len());
-                let name = &tail[..end];
-                // Only where it is actually read from the environment, which is
-                // what makes it a redirect rather than a mention.
-                let context_start = at.saturating_sub(40);
-                if rest[context_start..at].contains("env::var") {
-                    read_by_the_daemon.insert(name.to_owned());
-                }
-                rest = &tail[end..];
-            }
-        }
-        assert!(
-            !read_by_the_daemon.is_empty(),
-            "the scan found nothing, so it is not checking anything",
-        );
-
-        let handled: std::collections::BTreeSet<String> = DAEMON_REDIRECT_ENV
-            .iter()
-            .chain(DAEMON_INTENDED_ENV.iter())
-            .map(|name| (*name).to_owned())
-            .collect();
-        let unconsidered: Vec<&String> = read_by_the_daemon.difference(&handled).collect();
-        assert!(
-            unconsidered.is_empty(),
-            "the daemon reads these and this build neither strips nor \
-             deliberately passes them: {unconsidered:?}. Add each to \
-             DAEMON_REDIRECT_ENV, or to DAEMON_INTENDED_ENV if the app sets it \
-             on purpose.",
-        );
-
-        // And the stripping is actually wired, before the deliberate ones are
-        // set -- `env` after `env_remove` is what makes those still win.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let start = source
-            .find("fn spawn_locald()")
-            .expect("spawn_locald exists");
-        let body = &source[start..start + 3000];
-        let removed = body
-            .find("command.env_remove(name)")
-            .expect("a release build strips them");
-        let set = body
-            .find(r#".env("PATH", enriched_path())"#)
-            .expect("it sets PATH");
-        assert!(
-            removed < set,
-            "removal has to happen before the deliberate sets"
-        );
-    }
-
-    /// The dialogs name the machine this build actually runs on.
-    #[test]
-    fn native_copy_does_not_name_the_wrong_hardware() {
-        assert_eq!(
-            THIS_COMPUTER,
-            if cfg!(target_os = "windows") {
-                "this PC"
-            } else if cfg!(target_os = "macos") {
-                "this Mac"
-            } else {
-                "this computer"
-            }
-        );
-        let body = quit_prompt_body(&["Schedules stop.".into()]);
-        assert!(body.contains(THIS_COMPUTER), "{body}");
-
-        // The two highest-stakes strings in the app -- what is about to be
-        // deleted -- were the ones this test did not reach. Both shipped in the
-        // Windows build naming hardware that build's users do not have.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        for name in ["fn reset_local_data_impl(", "fn reset_full_reinstall_impl("] {
-            let start = source.find(name).unwrap_or_else(|| panic!("{name} exists"));
-            let body = &source[start..start + 1200];
-            assert!(
-                !body.contains("this Mac"),
-                "{name} hardcodes 'this Mac' in the copy that names what is deleted",
-            );
-            assert!(
-                body.contains("THIS_COMPUTER"),
-                "{name} must name the machine this build actually runs on",
-            );
-        }
-        // The other half of the bargain is stated where the decision is made.
-        assert!(
-            body.contains("To leave Lemma running, close the window instead."),
-            "{body}"
-        );
-    }
-
-    #[test]
-    fn a_window_is_only_restored_where_a_hand_can_reach_it() {
-        let screen = |x, y, w, h| {
-            (
-                tauri::PhysicalPosition::new(x, y),
-                tauri::PhysicalSize::new(w, h),
-            )
-        };
-        let at = |x, y| WindowPlacement {
-            position: tauri::PhysicalPosition::new(x, y),
-            size: tauri::PhysicalSize::new(1280, 860),
-        };
-        let laptop = [screen(0, 0, 1728, 1117)];
-        // The same desk, with the external display to the left -- which is
-        // where negative coordinates come from and why this cannot just clamp
-        // to zero.
-        let two_displays = [screen(0, 0, 1728, 1117), screen(-3440, -200, 3440, 1440)];
-
-        assert!(placement_is_reachable(&at(100, 100), &laptop));
-        assert!(placement_is_reachable(&at(-3000, 0), &two_displays));
-        assert!(
-            !placement_is_reachable(&at(-3000, 0), &laptop),
-            "the second display is gone; this window would be invisible"
-        );
-
-        // A window may legitimately hang off an edge. What must stay on screen
-        // is enough of the title bar to grab -- and "some overlap" is not that.
-        assert!(
-            placement_is_reachable(&at(1600, 20), &laptop),
-            "mostly off the right edge, but 128px of title bar is draggable"
-        );
-        assert!(
-            !placement_is_reachable(&at(1700, 20), &laptop),
-            "28px of chrome poking over the edge is not something a hand catches"
-        );
-        assert!(
-            !placement_is_reachable(&at(200, 1100), &laptop),
-            "the title bar is below the display, so there is nothing to drag"
-        );
-        assert!(
-            !placement_is_reachable(&at(200, -90), &laptop),
-            "the title bar is above the display"
-        );
-
-        // Asking the OS can fail, and a machine mid-display-change reports no
-        // monitors at all. Neither is evidence the saved value is wrong, and
-        // refusing to restore there would look like the bug this fixes.
-        assert!(placement_is_reachable(&at(100, 100), &[]));
-    }
-
-    #[test]
-    fn a_launch_tells_the_auth_portal_where_to_come_back_to() {
-        // The portal auto-continues an existing session only when it is given
-        // somewhere to go; without one it stops on a Continue button. Every
-        // cold start reached it without one, because a new runtime generation
-        // means the recorded resume target never matches.
-        let url = local_auth_url_returning_to("http://app.lemma.localhost:3711/", "signup", "/");
-        assert!(
-            url.starts_with("http://app.lemma.localhost:3711/auth?"),
-            "{url}"
-        );
-        assert!(url.contains("show=signup"), "{url}");
-        assert!(
-            url.contains("redirect_uri=%2F"),
-            "the portal needs a return address to continue without asking: {url}"
-        );
-
-        // Relative, so it resolves against the portal's own origin and cannot
-        // be aimed off it -- and so it survives locald allocating a different
-        // port than the one this launch used.
-        let sneaky = local_auth_url_returning_to(
-            "http://app.lemma.localhost:3711",
-            "signin",
-            "https://evil.example",
-        );
-        assert!(sneaky.contains("redirect_uri=%2F"), "{sneaky}");
-        assert!(!sneaky.contains("evil.example"), "{sneaky}");
-    }
-
-    #[test]
-    fn a_restored_window_comes_back_the_size_it_was_left() {
-        // The numbers are a real record from a 3024x1964 Retina display: the
-        // window filled the screen below the menu bar. Handed to the builder as
-        // written they mean 3024x1898 *logical* -- twice the screen -- so macOS
-        // clamped the size and put the window 66 points too low, which is what
-        // "it doesn't open full and the top is cut off" actually was.
-        let saved = WindowPlacement {
-            position: tauri::PhysicalPosition::new(0, 66),
-            size: tauri::PhysicalSize::new(3024, 1898),
-        };
-
-        let (position, size) = placement_in_logical(&saved, 2.0).expect("restorable");
-        assert_eq!((position.x, position.y), (0.0, 33.0));
-        assert_eq!((size.width, size.height), (1512.0, 949.0));
-
-        // A 1:1 display is the case that always worked, and must keep working.
-        let (position, size) = placement_in_logical(&saved, 1.0).expect("restorable");
-        assert_eq!((position.x, position.y), (0.0, 66.0));
-        assert_eq!((size.width, size.height), (3024.0, 1898.0));
-    }
-
-    #[test]
-    fn a_window_smaller_than_the_app_allows_is_not_restored() {
-        // 1200x800 physical is a legitimate record on a 1:1 screen and half the
-        // app's minimum on a 2x one. The floor is a logical size, so it can
-        // only be applied after the conversion -- applying it to the physical
-        // numbers is how a window half the allowed size gets restored and then
-        // clamped by the OS into a shape nobody chose.
-        let saved = WindowPlacement {
-            position: tauri::PhysicalPosition::new(0, 0),
-            size: tauri::PhysicalSize::new(1200, 800),
-        };
-        assert!(placement_in_logical(&saved, 1.0).is_some());
-        assert!(
-            placement_in_logical(&saved, 2.0).is_none(),
-            "600x400 logical is below the {}x{} minimum",
-            MIN_RESTORED.0,
-            MIN_RESTORED.1
-        );
-        // A monitor that reports nonsense must not produce an infinite window.
-        assert!(placement_in_logical(&saved, 0.0).is_none());
-        assert!(placement_in_logical(&saved, f64::NAN).is_none());
-    }
-
-    #[test]
-    fn a_placement_this_build_would_not_have_written_is_ignored() {
-        let saved = |value: serde_json::Value| saved_placement(&json!({ "window": value }));
-
-        assert_eq!(
-            saved(json!({"x": 12, "y": 34, "width": 1280, "height": 860})),
-            Some(WindowPlacement {
-                position: tauri::PhysicalPosition::new(12, 34),
-                size: tauri::PhysicalSize::new(1280, 860),
-            })
-        );
-        // Below `min_inner_size`: written by a build with different minimums,
-        // or captured mid-animation. The OS would clamp it and the window would
-        // come back a shape nobody chose.
-        assert_eq!(
-            saved(json!({"x": 0, "y": 0, "width": 400, "height": 300})),
-            None
-        );
-        // A truncated or hand-edited file must not stop the app opening.
-        assert_eq!(saved(json!({"x": 0, "y": 0, "width": 1280})), None);
-        assert_eq!(
-            saved(json!({"x": "left", "y": 0, "width": 1280, "height": 860})),
-            None
-        );
-        assert_eq!(saved(json!(null)), None);
-        assert_eq!(saved_placement(&json!({})), None, "a first-ever launch");
-        // Negative sizes and values past a u32 are refused rather than wrapped.
-        assert_eq!(
-            saved(json!({"x": 0, "y": 0, "width": -1280, "height": 860})),
-            None
-        );
-    }
-
-    /// A rebuild's own placement always wins over the remembered one.
-    ///
-    /// The two are different questions: a rebuild is "put it back exactly where
-    /// the user is looking", and a cold start is "open it where they left it
-    /// last time". Reading the remembered value first would move a window
-    /// during a server switch, which is the bug `WindowPlacement` was
-    /// introduced to fix.
-    #[test]
-    fn a_rebuild_keeps_the_window_where_it_is_rather_than_where_it_once_was() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        assert!(
-            source.contains("let placement = placement.or_else(|| remembered_placement(handle));"),
-            "the caller's placement has to take precedence",
-        );
-    }
-
-    /// Geometry is recorded as it changes, not only when the app is quit.
-    ///
-    /// An app that is force-killed, replaced by an updater, or caught in a
-    /// machine restart never sees a close event -- and those are exactly the
-    /// launches where coming back at the wrong size is most irritating.
-    #[test]
-    fn window_geometry_survives_a_launch_that_was_never_a_clean_quit() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let start = source
-            .find("fn build_main_window_at(")
-            .expect("build_main_window_at exists");
-        let body = &source[start..];
-        let moved = body
-            .find("tauri::WindowEvent::Moved(_) | tauri::WindowEvent::Resized(_)")
-            .expect("placement is recorded on move and resize");
-        assert!(
-            body[moved..moved + 200].contains("remember_placement(&window)"),
-            "both events have to write the placement"
-        );
-    }
-
-    #[test]
-    fn a_replacement_window_is_never_left_invisible() {
-        // Hidden-until-painted is only safe because something shows it anyway.
-        // A window that never paints and never appears is an app with no
-        // interface, which is the failure the label wait already guards.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let start = source
-            .find("fn build_main_window_at(")
-            .expect("build_main_window_at exists");
-        let body = &source[start..];
-
-        assert!(
-            body.contains("REPLACEMENT_REVEAL_TIMEOUT"),
-            "the reveal has a deadline"
-        );
-        let hidden = body
-            .find(".visible(false)")
-            .expect("a replacement starts hidden");
-        let backstop = body
-            .find("REPLACEMENT_REVEAL_TIMEOUT")
-            .expect("the deadline is used");
-        assert!(
-            hidden < backstop,
-            "hidden first, then the backstop that shows it"
-        );
-    }
-
-    #[test]
-    fn the_window_swap_waits_between_destroying_and_rebuilding() {
-        // The bug this pins: `destroy` returns as soon as the request is
-        // posted, so building immediately afterwards failed with `a webview
-        // with label `main` already exists` -- and because the fallback ran in
-        // the same millisecond it failed identically, leaving the app running
-        // with no window at all. Both build attempts have to be preceded by a
-        // wait.
-        //
-        // Asserted on the source because reaching the real path needs a running
-        // event loop; `wait_until_label_released` itself is tested above.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let body = {
-            let start = source
-                .find("fn rebuild_main_window_for_mode(app: &AppHandle, mode: &str)")
-                .expect("rebuild_main_window_for_mode exists");
-            let end = source[start..]
-                .find("\nfn ")
-                .map_or(source.len(), |offset| start + offset);
-            &source[start..end]
-        };
-
-        let destroy = body.find(".destroy()").expect("it destroys the old window");
-        let first_build = body
-            .find("build_main_window_at(app, mode, initial.clone(), true, true, placement)")
-            .expect("it rebuilds partitioned");
-        let fallback = body
-            .find("build_main_window_at(app, mode, initial, false, true, placement)")
-            .expect("it falls back to the shared store");
-        let waits: Vec<usize> = body
-            .match_indices("wait_until_label_released")
-            .map(|(index, _)| index)
-            .collect();
-
-        assert_eq!(waits.len(), 2, "every rebuild attempt waits for the label");
-        assert!(
-            destroy < waits[0] && waits[0] < first_build,
-            "the partitioned rebuild must wait after destroy"
-        );
-        assert!(
-            waits[1] < fallback,
-            "the shared-store fallback must wait too, or it repeats the failure \
-             it is supposed to recover from"
-        );
-    }
-
-    #[test]
-    fn choosing_a_connection_mode_re_enables_the_menus_it_gates() {
-        // Both menus gate their local-only verbs on `local`, and both are built
-        // during setup — which on a first launch is before anyone has chosen.
-        // Every gated item was created disabled and never revisited, so picking
-        // Local left "Local settings…" greyed out in the tray and Cmd-, dead in
-        // the app menu until Lemma was restarted.
-        //
-        // Asserted on the source because the alternative needs a running
-        // AppHandle, and the thing worth pinning is that the one function every
-        // mode change goes through is what rebuilds them.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let set_mode = {
-            let start = source
-                .find("fn set_mode(app: &AppHandle, mode: &str)")
-                .expect("set_mode exists");
-            let end = source[start..]
-                .find("\nfn ")
-                .map_or(source.len(), |offset| start + offset);
-            &source[start..end]
-        };
-        assert!(
-            set_mode.contains("refresh_menus_for_connection_mode(app)"),
-            "set_mode must rebuild the menus, or the choice does not take effect \
-             until the next launch"
-        );
-    }
-
-    #[test]
-    fn the_menu_bar_speaks_the_products_language() {
-        // There was no app menu at all before this: the shipped build used
-        // Tauri's default, so there was no Cmd-, and every Lemma verb was in
-        // the tray. The tray in turn read as a supervisor console.
-        // Scoped to the two menu builders rather than the whole file, because
-        // a test that scans its own source matches the very strings it is
-        // asserting are gone.
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let menus = {
-            let start = source
-                .find("fn build_app_menu")
-                .expect("the app menu builder exists");
-            let end = source
-                .find("fn disconnect_locald")
-                .expect("tray builder ends");
-            &source[start..end]
-        };
-
-        assert!(menus
-            .split_whitespace()
-            .collect::<String>()
-            .contains("\"Desktopsettings…\",true,Some(\"CmdOrCtrl+,\")"));
-        assert!(menus.contains("\"Connection…\""));
-
-        // Quit is one command, and it is the one that stops the local server.
-        // There used to be two ways to leave — Quit, which left the VM and the
-        // whole stack running with no owner on screen, and a "Stop the local
-        // server and quit" buried in Troubleshoot that did what people mean by
-        // quitting. Closing the window already covers "go away and keep
-        // serving", and keeps the tray as a way back, so the middle state was
-        // strictly worse than both neighbours.
-        assert!(
-            !menus.contains("quit-and-stop") && !menus.contains("Stop the local server and quit"),
-            "there is exactly one Quit, and it stops the local server",
-        );
-        assert!(menus.contains("\"Stop the local server\""));
-
-        // ⌘Q has to reach `request_quit`, which means owning the item. AppKit's
-        // predefined quit is `terminate:` and cannot be intercepted, so it would
-        // silently skip the prompt and take the stack down — or leave it up —
-        // without asking. Matched as a call, because this test reads its own
-        // source and the comment above the item names the thing it rules out.
-        assert!(
-            menus.contains("\"quit\", \"Quit Lemma\", true, Some(\"CmdOrCtrl+Q\")"),
-            "Quit is an app-owned item bound to CmdOrCtrl+Q",
-        );
-        assert!(
-            !menus.contains("PredefinedMenuItem::quit("),
-            "a predefined quit cannot be asked about first",
-        );
-
-        // The operator vocabulary this replaced must not come back.
-        for retired in [
-            "Stop Services and Infra",
-            "Switch Connection Mode",
-            "Start Services",
-        ] {
-            assert!(
-                !menus.contains(retired),
-                "{retired:?} is supervisor vocabulary, not a product menu item"
-            );
-        }
-    }
-
-    #[test]
-    fn quit_watchdog_outlives_the_owned_runtime_cleanup_deadlines() {
-        let cleanup = RELEASE_ON_EXIT_TIMEOUT
-            + LOCALD_HANDSHAKE_BUDGET * 2
-            + LOCALD_EXIT_POLL * (QUIT_DAEMON_GRACE_ATTEMPTS + LOCALD_FORCE_EXIT_ATTEMPTS) as u32
-            + VM_STOP_GRACE_BUDGET
-            + VM_STOP_REAP_BUDGET;
-        assert!(
-            QUIT_DAEMON_BUDGET > cleanup + Duration::from_secs(5),
-            "the shell must not terminate its cleanup worker before it can stop an unresponsive VM"
-        );
-    }
-
-    #[test]
-    fn quitting_names_the_work_it_is_about_to_stop() {
-        // The whole point of the prompt is that none of this is on screen. A
-        // warning that says "are you sure?" and nothing else would be worse than
-        // no warning, because it teaches people to dismiss it.
-        let running_host =
-            json!({"running": true, "targets": [{"name": "work"}, {"name": "home"}]});
-        let lines = quit_impact_lines(true, Some(&running_host), Some("public"));
-        assert_eq!(
-            lines,
-            vec![
-                "Schedules and background work stop running.",
-                "The agents on this computer stop answering (2 paired workspaces).",
-                "Your public link closes.",
-            ]
-        );
-
-        // Singular reads as English, and an enabled-but-unpaired host still
-        // stops answering, so it is still worth one line.
-        let unpaired = json!({"running": true, "targets": []});
-        assert_eq!(
-            quit_impact_lines(false, Some(&unpaired), None),
-            vec!["The agents on this computer stop answering."]
-        );
-        let one = json!({"running": true, "targets": [{"name": "work"}]});
-        assert!(quit_impact_lines(false, Some(&one), None)[0].ends_with("(1 paired workspace)."));
-
-        assert_eq!(
-            quit_impact_lines(true, None, Some("local_network")),
-            vec![
-                "Schedules and background work stop running.",
-                "Your local network link closes.",
-            ]
-        );
-    }
-
-    #[test]
-    fn a_quit_with_nothing_running_asks_nothing() {
-        // Every dialog on the way out has to earn itself. A stopped stack with
-        // no Agent Host and no shared link costs the user nothing to quit, and
-        // being asked anyway is how a prompt becomes noise.
-        let idle_host = json!({"running": false, "targets": []});
-        assert!(quit_impact_lines(false, Some(&idle_host), Some("this_computer")).is_empty());
-        assert!(quit_impact_lines(false, None, None).is_empty());
-    }
-
-    /// A hosted user with a running Agent Host is warned too.
-    ///
-    /// `quit_impact` used to return empty for any non-local mode, so quitting
-    /// asked nothing at all. But locald is started in hosted mode *precisely*
-    /// so the Agent Host can run, and a full quit stops it -- so somebody with
-    /// a coding agent mid-run lost it silently, while a local user got a
-    /// careful three-line warning. Only the stack line is local-only.
-    #[test]
-    fn a_hosted_quit_still_names_a_running_agent_host() {
-        let running = json!({"running": true, "targets": ["workspace-a"]});
-        let hosted = quit_impact_lines(/* stack_up */ false, Some(&running), None);
-        assert_eq!(hosted.len(), 1, "{hosted:?}");
-        assert!(hosted[0].contains("agents on this computer"), "{hosted:?}");
-        assert!(
-            !hosted.iter().any(|line| line.contains("Schedules")),
-            "a hosted workspace has no local stack to stop: {hosted:?}",
-        );
-    }
-
-    /// A confirmed quit is never left waiting forever on a stop.
-    ///
-    /// `quit_after_stop` is consumed only by a `done` event saying the stop
-    /// succeeded, so a wedged VM left the app running on "Winding down." with
-    /// the user's quit unanswered -- and the error screen's button read "Try
-    /// again", offering to *start* Lemma to somebody who had asked to leave.
-    /// The shell's view of the daemon, folded one event at a time.
-    ///
-    /// `handle_locald_event` was 391 lines with no test at all — the path that
-    /// decides what every screen shows, checked only by running the app. These
-    /// drive the reducer it was split into.
-    mod locald_events {
-        use super::super::{apply_locald_event, UiState};
-        use serde_json::json;
-
-        /// A phase event is progress, and progress is not an error.
-        #[test]
-        fn a_phase_event_describes_the_work_without_claiming_readiness() {
-            let mut ui = UiState {
-                ready: true,
-                ..UiState::default()
-            };
-
-            apply_locald_event(
-                &mut ui,
-                "phase",
-                &json!({
-                    "label": "Starting authentication",
-                    "key": "supertokens",
-                    "progress": 40,
-                    "detail": "waiting for the guest",
-                }),
-            );
-
-            assert_eq!(ui.phase_key, "supertokens");
-            assert_eq!(ui.progress, 40);
-            assert_eq!(ui.status, "Starting authentication: waiting for the guest");
-            assert!(!ui.ready, "work in progress is not a ready workspace");
-            assert!(!ui.error);
-        }
-
-        /// Lifecycle state outranks a stale phase.
-        ///
-        /// Older daemons legitimately report `stopped` while their last phase
-        /// still reads ready at 100%. Taking the phase at face value showed
-        /// "It's ready" over a stack that had stopped.
-        #[test]
-        fn a_stopped_daemon_is_shown_as_stopped_even_if_its_last_phase_said_ready() {
-            let mut ui = UiState {
-                phase: "Ready".into(),
-                phase_key: "ready".into(),
-                progress: 100,
-                ready: true,
-                ..UiState::default()
-            };
-
-            apply_locald_event(
-                &mut ui,
-                "state",
-                &json!({"running": false, "ready": false, "status": "stopped"}),
-            );
-
-            assert_eq!(ui.phase_key, "stopped");
-            assert_eq!(ui.progress, 0);
-            assert!(!ui.ready);
-        }
-
-        /// An error the user can act on survives the ordinary status traffic
-        /// that follows it, or the screen offering the fix disappears before it
-        /// can be read.
-        #[test]
-        fn an_actionable_error_is_not_cleared_by_the_next_status() {
-            let mut ui = UiState {
-                error: true,
-                error_code: "wsl-required".into(),
-                ..UiState::default()
-            };
-
-            apply_locald_event(
-                &mut ui,
-                "status",
-                &json!({"running": false, "ready": false, "status": "idle"}),
-            );
-
-            assert!(ui.error, "the recovery screen must not vanish on its own");
-            assert_eq!(ui.error_code, "wsl-required");
-        }
-
-        /// Reaching ready is what clears it.
-        #[test]
-        fn becoming_ready_clears_a_previous_error() {
-            let mut ui = UiState {
-                error: true,
-                error_code: "locald-start-failed".into(),
-                ..UiState::default()
-            };
-
-            apply_locald_event(
-                &mut ui,
-                "status",
-                &json!({"running": true, "ready": true, "status": "ready"}),
-            );
-
-            assert!(!ui.error);
-            assert!(ui.error_code.is_empty());
-            assert!(ui.ready);
-        }
-
-        /// Recovery options are fetched once per error, not on every event that
-        /// repeats it — the daemon reports status continuously while stopped.
-        #[test]
-        fn terminal_recovery_is_scheduled_once_for_one_error() {
-            let mut ui = UiState::default();
-            let failure = json!({"running": false, "ready": false, "status": "error", "code": "x"});
-
-            let first = apply_locald_event(&mut ui, "state", &failure);
-            assert!(first.schedule_terminal_recovery, "the first error asks");
-
-            let second = apply_locald_event(&mut ui, "status", &failure);
-            assert!(
-                !second.schedule_terminal_recovery,
-                "repeats of the same error must not ask again"
-            );
-        }
-
-        /// A workspace URL is adopted only when both halves are trusted.
-        #[test]
-        fn an_untrusted_workspace_url_is_refused() {
-            let mut ui = UiState::default();
-
-            apply_locald_event(
-                &mut ui,
-                "status",
-                &json!({
-                    "running": true,
-                    "ready": true,
-                    "status": "ready",
-                    "url": "https://evil.example",
-                    "api_url": "https://evil.example/api",
-                }),
-            );
-
-            assert!(
-                ui.url.is_empty(),
-                "the shell must not navigate anywhere the daemon names: {}",
-                ui.url
-            );
-        }
-    }
-
-    /// Each platform's data disk, spelled the way that platform spells it.
-    ///
-    /// The Windows branch pointed at `runtime/windows`, which nothing creates.
-    /// So on Windows `has_local_runtime_data` answered "no data" for a real
-    /// installation — the guard that refuses to reset a user's data during an
-    /// update was leaning entirely on the config check — and Recovery reported
-    /// the disk as zero bytes.
-    #[test]
-    fn the_managed_data_disk_is_the_one_this_platform_actually_writes() {
-        let disk = managed_data_disk();
-        let tail: Vec<_> = disk
-            .components()
-            .rev()
-            .take(2)
-            .map(|part| part.as_os_str().to_string_lossy().into_owned())
-            .collect();
-
-        if cfg!(windows) {
-            assert_eq!(tail, vec!["ext4.vhdx", "wsl"], "{}", disk.display());
-        } else {
-            assert_eq!(tail, vec!["data.raw", "macos"], "{}", disk.display());
-        }
-        assert!(
-            disk.starts_with(locald_root()),
-            "the disk belongs to this installation: {}",
-            disk.display()
-        );
-    }
-
-    /// A build keeps its data where its own name says, not where "Lemma" does.
-    ///
-    /// Qualifying a candidate means running it on the same Mac as the real
-    /// installation. Sharing `Application Support/Lemma` would let the
-    /// candidate stop the user's daemon, adopt its runtime and reset its pods
-    /// -- so a QA build bakes in a different directory, and a release build
-    /// must keep the one every existing installation already uses.
-    #[test]
-    fn a_release_build_keeps_its_data_where_installed_lemma_already_has_it() {
-        assert_eq!(
-            option_env!("LEMMA_DESKTOP_DATA_DIR_NAME"),
-            None,
-            "a build with this set is a qualification candidate, not a release"
-        );
-        assert_eq!(DATA_DIR_NAME, "Lemma");
-        assert!(
-            app_support_dir().ends_with("Lemma"),
-            "moving this orphans every existing installation's data: {}",
-            app_support_dir().display()
-        );
-    }
-
-    /// An update that fails after the stack is stopped must say the previous
-    /// version survived, and whether it is running.
-    ///
-    /// `install_app_update` stops locald before installing, on purpose: an
-    /// in-place update writes to the same path, so a stale daemon would be
-    /// adopted by the new shell. But when the install then failed it returned
-    /// the error and stopped there, leaving someone in Local settings looking
-    /// at a workspace whose backend had gone, with nothing on screen offering
-    /// to bring it back.
-    #[test]
-    fn a_failed_update_says_the_previous_version_survived_it() {
-        let recovered = failed_install_message("the bundle is not writable", None);
-        assert!(
-            recovered.contains("previous version is still installed"),
-            "{recovered}"
-        );
-        assert!(
-            recovered.contains("starting again"),
-            "the user needs to know the stack is coming back: {recovered}"
-        );
-
-        let stranded = failed_install_message(
-            "the bundle is not writable",
-            Some("daemon did not answer".into()),
-        );
-        assert!(
-            stranded.contains("previous version is still installed"),
-            "{stranded}"
-        );
-        assert!(
-            stranded.contains("Recovery"),
-            "a stack that stayed down has to name the way out: {stranded}"
-        );
-        assert!(
-            stranded.contains("daemon did not answer"),
-            "the reason it stayed down is the actionable half: {stranded}"
-        );
-    }
-
-    /// Choosing "Keep waiting" must bring the offer back, not retire it.
-    ///
-    /// The watchdog used to ask once. On "Keep waiting" it re-armed the flag
-    /// and returned, so a stop that never confirmed left the app on "Winding
-    /// down." for ever with no way out -- repeating the shortcut does not help,
-    /// because `request_quit` returns early once the quit is confirmed.
-    #[test]
-    fn a_confirmed_quit_keeps_offering_to_leave_until_the_stop_finishes() {
-        let asked = std::cell::Cell::new(0);
-        let rearmed = std::cell::Cell::new(0);
-        let left = std::cell::Cell::new(false);
-
-        run_quit_watchdog(
-            || {},
-            || true, // The stop never lands.
-            || {
-                asked.set(asked.get() + 1);
-                // Wait three times, then take the way out.
-                asked.get() > 3
-            },
-            || rearmed.set(rearmed.get() + 1),
-            || left.set(true),
-        );
-
-        assert_eq!(asked.get(), 4, "each budget must ask again");
-        assert_eq!(rearmed.get(), 3, "waiting must re-arm so a late stop quits");
-        assert!(left.get(), "Quit Anyway must actually leave");
-    }
-
-    /// The other exit: the stop lands, and nobody is asked anything.
-    #[test]
-    fn a_quit_watchdog_stands_down_once_the_stop_finishes() {
-        let asked = std::cell::Cell::new(0);
-        let left = std::cell::Cell::new(false);
-
-        run_quit_watchdog(
-            || {},
-            || false, // `quit_after_stop` was consumed by a `done` event.
-            || {
-                asked.set(asked.get() + 1);
-                true
-            },
-            || {},
-            || left.set(true),
-        );
-
-        assert_eq!(asked.get(), 0, "a finished stop must not prompt");
-        assert!(!left.get(), "the quit already completed on its own");
-    }
-
-    /// A web inspector does not ship enabled in the top-level menus.
-    #[test]
-    fn developer_tools_are_a_development_build_affordance() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        for block in source.split("\"devtools\",").skip(1) {
-            let head = &block[..block.len().min(200)];
-            assert!(
-                head.contains("cfg!(debug_assertions)"),
-                "a release build must not offer a web inspector into a webview \
-                 that talks to the workspace over IPC: {head}",
-            );
-        }
-    }
-
-    #[test]
-    fn the_quit_prompt_offers_the_alternative_it_is_replacing() {
-        // Someone pressing ⌘Q may mean "get out of my way", which is what
-        // closing the window does — and unlike this, it keeps everything
-        // serving. The prompt has to say so, or the only discoverable way to
-        // keep schedules running is to already know about it.
-        let body = quit_prompt_body(&["Schedules and background work stop running.".into()]);
-        assert!(body.contains("Schedules and background work stop running."));
-        assert!(body.contains("close the window"));
-        // And it has to say what is not lost, or "stop" reads as "delete".
-        assert!(body.contains(&format!("stay on {THIS_COMPUTER}")));
-    }
-
-    #[test]
-    fn the_window_layer_is_painted_in_both_appearances() {
-        // Whatever the webview is not painting, this is. Neither may be the
-        // macOS default of nothing, which composites black.
-        assert_eq!(canvas_color(tauri::Theme::Dark), CANVAS_DARK);
-        assert_eq!(canvas_color(tauri::Theme::Light), CANVAS_LIGHT);
-    }
-
-    #[test]
-    fn a_resume_opens_the_recorded_route_rather_than_the_workspace_root() {
-        // Opening the root means loading the app once to authenticate and
-        // resolve the last pod, then loading it again at the pod it resolved
-        // to. The recorded route skips the first of those.
-        let target = ResumeTarget {
-            url: "http://app.lemma.localhost:49180".into(),
-            api_url: "http://app.lemma.localhost:49181".into(),
-            generation: "abc123".into(),
-            release: env!("CARGO_PKG_VERSION").into(),
-            route: "/pod/42/conversations/7".into(),
-        };
-        assert_eq!(
-            resume_entry_url(&target),
-            "http://app.lemma.localhost:49180/pod/42/conversations/7"
-        );
-
-        // An install that has only ever seen the root still resumes; it just
-        // resumes at the root, without a doubled slash.
-        let root = ResumeTarget {
-            route: "/".into(),
-            ..target
-        };
-        assert_eq!(resume_entry_url(&root), "http://app.lemma.localhost:49180");
-    }
-
-    #[test]
-    fn a_resume_target_from_another_release_is_refused() {
-        // The failure this prevents: installing a new build over an old one,
-        // whose stack is often still serving. The probe passes, the window
-        // opens the old workspace, and then ensure_locald replaces the
-        // mismatched daemon and brings everything back on new ports — leaving
-        // the window on a dead port with nothing to navigate it away.
-        let saved = |release: &str| {
-            json!({
-                "url": "http://app.lemma.localhost:49180",
-                "apiUrl": "http://app.lemma.localhost:49181",
-                "generation": "abc123",
-                "release": release,
-                "route": "/",
-            })
-        };
-        let accepted = |entry: &Value| {
-            entry["release"].as_str() == Some(env!("CARGO_PKG_VERSION"))
-                && !entry["generation"].as_str().unwrap_or_default().is_empty()
-        };
-
-        assert!(accepted(&saved(env!("CARGO_PKG_VERSION"))));
-        assert!(!accepted(&saved("0.6.9")));
-        // A target written before releases were recorded has no claim either.
-        assert!(!accepted(&json!({
-            "url": "http://app.lemma.localhost:49180",
-            "apiUrl": "http://app.lemma.localhost:49181",
-            "generation": "abc123",
-            "route": "/",
-        })));
-    }
-
-    #[test]
-    fn a_ready_event_rescues_a_window_left_on_a_stale_workspace() {
-        // Three cases, and the middle one is the bug: the window is showing a
-        // workspace, just not this one. The old test was "is it the splash?",
-        // which answered no and left it there.
-        let splash = tauri::Url::parse(&native_asset_url("index.html")).unwrap();
-        let current = tauri::Url::parse("http://app.lemma.localhost:49180/pod/7").unwrap();
-        let stale = tauri::Url::parse("http://app.lemma.localhost:57919/pod/7").unwrap();
-        let workspace = "http://app.lemma.localhost:49180";
-
-        let needs = |url: &tauri::Url| native_splash_url(url) || !same_origin(url, workspace);
-
-        assert!(needs(&splash), "the ordinary first start");
-        assert!(needs(&stale), "a resume whose stack was replaced under it");
-        assert!(
-            !needs(&current),
-            "a stable workspace must never be navigated for a transient event",
-        );
-    }
-
-    #[test]
-    fn a_resume_target_is_only_trusted_with_a_generation_and_workspace_origins() {
-        // The generation is what separates "our stack is still serving" from
-        // "something is listening on that port". Without it, a resume would
-        // hand the window to whatever answered.
-        let saved = |generation: &str, url: &str| {
-            json!({
-                "resumeTarget": {
-                    "url": url,
-                    "apiUrl": "http://app.lemma.localhost:49181",
-                    "generation": generation,
-                    "route": "/",
-                }
-            })
-        };
-        let parse = |value: Value| -> bool {
-            let entry = &value["resumeTarget"];
-            let generation = entry["generation"].as_str().unwrap_or_default();
-            !generation.is_empty()
-                && trusted_workspace_urls(
-                    entry["url"].as_str().unwrap_or_default(),
-                    entry["apiUrl"].as_str().unwrap_or_default(),
-                )
-        };
-
-        assert!(parse(saved("abc123", "http://app.lemma.localhost:49180")));
-        assert!(!parse(saved("", "http://app.lemma.localhost:49180")));
-        assert!(!parse(saved("abc123", "https://evil.example.com")));
-    }
-
-    #[test]
-    fn privileged_configuration_commands_are_control_window_only() {
-        assert!(is_control_window_label("control"));
-        assert!(!is_control_window_label("main"));
-        assert!(!is_control_window_label("sales.apps.lemma.localhost"));
-    }
-
-    #[test]
-    fn configured_origins_are_exact() {
-        let same = tauri::Url::parse("https://lemma.work/docs").unwrap();
-        let subdomain = tauri::Url::parse("https://untrusted.lemma.work/").unwrap();
-        let wrong_port = tauri::Url::parse("http://localhost:9999/").unwrap();
-
-        assert!(same_origin(&same, "https://lemma.work"));
-        assert!(!same_origin(&subdomain, "https://lemma.work"));
-        assert!(!same_origin(&wrong_port, "http://localhost:3711"));
-    }
-
-    #[test]
-    fn durable_daemon_must_match_the_bundled_host_pack_release() {
-        let root = tempfile::tempdir().unwrap();
-        let pack = root.path().join("local-runtime");
-        std::fs::create_dir_all(&pack).unwrap();
-        let current = json!({
-            "event": "hello", "protocol": 1, "mode": "host-packs",
-            "daemon_api_revision": REQUIRED_LOCALD_API_REVISION,
-            "host_pack_release": "1.2.3",
-            "host_pack_root": path_identity(&pack),
-        });
-        let old_same_version = json!({
-            "event": "hello", "protocol": 1, "mode": "host-packs",
-            "daemon_api_revision": REQUIRED_LOCALD_API_REVISION,
-            "host_pack_release": "1.2.3",
-        });
-        let stale_daemon = json!({
-            "event": "hello", "protocol": 1, "mode": "host-packs",
-            "host_pack_release": "1.2.3",
-            "host_pack_root": path_identity(&pack),
-        });
-        let compatibility = json!({
-            "event": "hello", "protocol": 1, "mode": "compatibility",
-        });
-
-        assert!(locald_matches_host_pack(
-            &current,
-            Some("1.2.3"),
-            Some(&pack)
-        ));
-        assert!(!locald_matches_host_pack(
-            &current,
-            Some("1.2.4"),
-            Some(&pack)
-        ));
-        assert!(!locald_matches_host_pack(
-            &old_same_version,
-            Some("1.2.3"),
-            Some(&pack)
-        ));
-        assert!(!locald_matches_host_pack(
-            &stale_daemon,
-            Some("1.2.3"),
-            Some(&pack)
-        ));
-        assert!(!locald_matches_host_pack(
-            &compatibility,
-            Some("1.2.3"),
-            Some(&pack)
-        ));
-        assert!(locald_matches_host_pack(&compatibility, None, None));
-    }
-
-    #[test]
-    fn unpublished_online_runtime_error_is_actionable_and_logged_in_app() {
-        // "Actionable" now means actionable *by the person reading it*. This
-        // used to assert the previous copy -- "publish its runtime artifacts,
-        // or use the compressed PR test DMG for this exact commit" -- which is
-        // an instruction to a maintainer that shipped to strangers.
-        let message = actionable_runtime_install_error(
-            "could not install local runtime: artifact download failed with HTTP 404",
-        );
-        assert!(message.contains("superseded"), "{message}");
-        assert!(message.contains("download the current one"), "{message}");
-        assert!(!message.contains("Publish"), "{message}");
-        assert!(!message.contains("PR test DMG"), "{message}");
-
-        let splash = include_str!("../ui/index.html").replace("\r\n", "\n");
-        assert!(splash.contains("diagnosticLogs: (source, cursor = null)"));
-        assert!(splash.contains("refreshDiagnosticLog"));
-        assert!(splash.contains("id=\"log-tabs\""));
-        assert!(splash.contains("View log"));
-    }
-
-    #[test]
-    fn ordinary_web_navigation_stays_in_the_webview() {
-        let urls = [
-            "https://sales.apps.lemma.work/",
-            "https://api.lemma.work/widgets/serve/conversation/tool",
-            "http://sales.apps.lemma.localhost:8711/",
-            "https://widgets.example.com/report",
-        ];
-
-        for raw_url in urls {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert_eq!(
-                navigation_disposition(&url, "hosted", "https://lemma.work", ""),
-                NavigationDisposition::Allow
-            );
-        }
-    }
-
-    #[test]
-    fn local_navigation_accepts_only_locald_owned_loopback_origins() {
-        let app_base = "http://app.lemma.localhost:63844";
-        let api_base = "http://app.lemma.localhost:63845";
-        for raw_url in [
-            "http://app.lemma.localhost:63844/auth",
-            "http://app.lemma.localhost:63845/files/download",
-            "http://sales.apps.lemma.localhost:63845/",
-        ] {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert_eq!(
-                navigation_disposition(&url, "local", app_base, api_base),
-                NavigationDisposition::Allow
-            );
-        }
-
-        for raw_url in [
-            "http://app.lemma.localhost:3710/",
-            "http://app.lemma.localhost:3711/verify-email",
-            "http://app.lemma.localhost:8710/",
-            "http://127.0.0.1:3000/",
-            "http://192.168.1.20:8000/",
-        ] {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert_eq!(
-                navigation_disposition(&url, "local", app_base, api_base),
-                NavigationDisposition::Deny
-            );
-        }
-    }
-
-    #[test]
-    fn iframes_may_render_their_own_inline_content() {
-        let app_base = "http://app.lemma.localhost:63844";
-        let api_base = "http://app.lemma.localhost:63845";
-
-        // The document viewer previews HTML and .docx through `srcdoc`, and the
-        // navigation delegate is asked about subframes too. Denying these blanked
-        // the preview.
-        for raw_url in ["about:srcdoc", "about:blank"] {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert_eq!(
-                navigation_disposition(&url, "local", app_base, api_base),
-                NavigationDisposition::Allow
-            );
-        }
-
-        // Nothing else off the http/https path comes along for the ride.
-        for raw_url in [
-            "data:text/html,<h1>x</h1>",
-            "file:///etc/passwd",
-            "about:settings",
-            "javascript:alert(1)",
-        ] {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert_eq!(
-                navigation_disposition(&url, "local", app_base, api_base),
-                NavigationDisposition::Deny
-            );
-        }
-    }
-
-    #[test]
-    fn downloads_are_judged_by_the_origin_that_minted_them() {
-        let app_base = "http://app.lemma.localhost:63844";
-        let api_base = "http://app.lemma.localhost:63845";
-
-        // Object URLs from the workspace itself: every Download button in the app.
-        for raw_url in [
-            "blob:http://app.lemma.localhost:63844/9f1c-uuid",
-            "blob:http://app.lemma.localhost:63845/9f1c-uuid",
-            // Bundle export navigates straight to a Content-Disposition endpoint.
-            "http://app.lemma.localhost:63845/pods/demo/bundle/download",
-        ] {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert!(
-                download_disposition(&url, "local", app_base, api_base),
-                "{raw_url} should download"
-            );
-        }
-
-        for raw_url in [
-            // A loopback origin this install does not own.
-            "blob:http://app.lemma.localhost:3710/9f1c-uuid",
-            "http://127.0.0.1:3000/export.csv",
-            // Schemes that should never reach the disk by themselves.
-            "file:///etc/passwd",
-            "data:text/csv,a%2Cb",
-            "blob:not-a-url",
-        ] {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert!(
-                !download_disposition(&url, "local", app_base, api_base),
-                "{raw_url} should not download"
-            );
-        }
-    }
-
-    #[test]
-    fn local_workspace_origin_requires_isolated_locald_ports_or_a_canonical_gateway() {
-        assert!(trusted_workspace_urls(
-            "http://app.lemma.localhost:63844",
-            "http://app.lemma.localhost:63845"
-        ));
-        assert!(trusted_workspace_urls(
-            "http://192.168.1.20:51324",
-            "http://192.168.1.20:51324/_lemma/api"
-        ));
-        assert!(trusted_workspace_urls(
-            "https://lemma-example.ngrok.app",
-            "https://lemma-example.ngrok.app/_lemma/api"
-        ));
-        assert!(!trusted_workspace_urls(
-            "http://app.lemma.localhost:3711",
-            "http://app.lemma.localhost:8711"
-        ));
-        assert!(!trusted_workspace_urls(
-            "http://app.lemma.localhost:63844",
-            "http://app.lemma.localhost:8710"
-        ));
-    }
-
-    #[test]
-    fn desktop_frontend_launcher_has_no_shared_development_origin_fallback() {
-        let launcher = include_str!("../runtime/frontend-launcher.mjs").replace("\r\n", "\n");
-        assert!(launcher.contains("locald must provide the isolated frontend and API origins"));
-        assert!(!launcher.contains("app.lemma.localhost:3711"));
-        assert!(!launcher.contains("app.lemma.localhost:8711"));
-    }
-
-    #[test]
-    fn unsupported_navigation_schemes_are_denied() {
-        for raw_url in [
-            "file:///tmp/report.html",
-            "javascript:alert(1)",
-            "lemma://other",
-        ] {
-            let url = tauri::Url::parse(raw_url).unwrap();
-            assert_eq!(
-                navigation_disposition(&url, "hosted", "https://lemma.work", ""),
-                NavigationDisposition::Deny
-            );
-        }
-    }
-
-    #[test]
-    fn explicit_new_windows_keep_the_browser_policy() {
-        let app_base = "https://lemma.work";
-        let first_party = tauri::Url::parse("https://lemma.work/docs").unwrap();
-        let external = tauri::Url::parse("https://widgets.example.com/report").unwrap();
-        let blank = tauri::Url::parse("about:blank").unwrap();
-
-        assert_eq!(
-            new_window_disposition(&first_party, "hosted", app_base, ""),
-            NewWindowDisposition::NavigateInApp
-        );
-        assert_eq!(
-            new_window_disposition(&external, "hosted", app_base, ""),
-            NewWindowDisposition::OpenExternal
-        );
-        assert_eq!(
-            new_window_disposition(&blank, "hosted", app_base, ""),
-            NewWindowDisposition::Deny
-        );
-    }
-
-    /// "Open in new window" on a pod app must not eat the window Lemma is in.
-    ///
-    /// It used to answer `NavigateInApp`, which pointed the *main* webview at
-    /// the app: the workspace, the sidebar and every tab were replaced by
-    /// somebody's dashboard, and nothing on screen led back. The only exit was
-    /// quitting the app.
-    #[test]
-    fn a_pod_app_opens_beside_lemma_rather_than_on_top_of_it() {
-        let api_base = "http://app.lemma.localhost:8711";
-        let app_base = "http://app.lemma.localhost:3000";
-        let pod_app = tauri::Url::parse("http://study-lab.apps.lemma.localhost:8711/").unwrap();
-
-        assert_eq!(
-            new_window_disposition(&pod_app, "local", app_base, api_base),
-            NewWindowDisposition::OpenAppWindow
-        );
-
-        // The workspace itself still belongs in the main window -- this is a
-        // rule about apps, not a general retreat from in-app navigation.
-        let workspace = tauri::Url::parse("http://app.lemma.localhost:3000/pods").unwrap();
-        assert_eq!(
-            new_window_disposition(&workspace, "local", app_base, api_base),
-            NewWindowDisposition::NavigateInApp
-        );
-
-        // And an app on a port that is not the API's is not our app: it must
-        // not inherit a window that skips the browser policy.
-        let impostor = tauri::Url::parse("http://study-lab.apps.lemma.localhost:9999/").unwrap();
-        assert_ne!(
-            new_window_disposition(&impostor, "local", app_base, api_base),
-            NewWindowDisposition::OpenAppWindow
-        );
-    }
-
-    /// A pod app window is a window, not a viewer: it can download.
-    ///
-    /// Tauri cancels a download outright when a webview registers no policy,
-    /// silently -- the failure this repo already hit once, where every Download
-    /// button on macOS did nothing and said nothing. An app that exports a CSV
-    /// is an ordinary app, and it opens in this window now.
-    #[test]
-    fn the_pod_app_window_can_download_what_an_app_offers() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let start = source
-            .find("fn open_pod_app_window(")
-            .expect("the app window builder exists");
-        let end = source[start..]
-            .find("\nfn ")
-            .map_or(source.len(), |offset| start + offset);
-        let builder = &source[start..end];
-        assert!(
-            builder.contains(".on_download("),
-            "the app window registers no download policy, so downloads from an \
-             app are cancelled with no error",
-        );
-        assert!(
-            builder.contains("download_disposition"),
-            "the app window must be held to the same download policy as the \
-             workspace, not a looser one of its own",
-        );
-    }
-
-    /// Closing the workspace must not strand a pod app window.
-    ///
-    /// Dropping to `Accessory` removes the Dock tile for the whole application,
-    /// so doing it while an app window is still on screen leaves that window
-    /// visible, un-⌘-tabbable, and owned by an app the Dock says is not running.
-    #[test]
-    fn the_dock_follows_what_is_actually_on_screen() {
-        let source = include_str!("main.rs").replace("\r\n", "\n");
-        let start = source
-            .find("fn settle_dock_presence(")
-            .expect("the dock helper exists");
-        let end = source[start..]
-            .find("\nfn ")
-            .map_or(source.len(), |offset| start + offset);
-        let helper = &source[start..end];
-        assert!(
-            helper.contains("windows()") && helper.contains("is_visible"),
-            "dock presence must be decided by what is visible, not by the \
-             workspace window alone",
-        );
-        assert!(
-            helper.contains("ActivationPolicy::Regular"),
-            "something on screen has to put the Dock icon back",
-        );
-    }
-
-    /// The app window is deliberately outside every capability file.
-    ///
-    /// Capabilities are scoped by webview label, so a label with no entry gets
-    /// no Tauri command surface at all. That is the whole reason a pod app can
-    /// have a window: it runs user-authored code, and an entry here would hand
-    /// it the IPC bridge.
-    #[test]
-    fn the_pod_app_window_is_granted_no_commands() {
-        for (name, source) in [
-            ("main", include_str!("../capabilities/main.json")),
-            ("control", include_str!("../capabilities/control.json")),
-            ("workspace", include_str!("../capabilities/workspace.json")),
-        ] {
-            assert!(
-                !source.contains(POD_APP_WINDOW),
-                "{name}.json names the pod-app window, which would give \
-                 user-authored app code a command surface"
-            );
-        }
-    }
-
-    #[test]
-    fn desktop_browser_login_is_explicitly_marked() {
-        let desktop = tauri::Url::parse(
-            "https://lemma.work/auth?desktop_browser=1&desktop_request=request-1234567890",
-        )
-        .unwrap();
-        let ordinary = tauri::Url::parse("https://lemma.work/auth").unwrap();
-        let unrelated = tauri::Url::parse("https://lemma.work/docs?desktop_browser=1").unwrap();
-
-        assert!(is_desktop_browser_auth_url(&desktop));
-        assert!(!is_desktop_browser_auth_url(&ordinary));
-        assert!(!is_desktop_browser_auth_url(&unrelated));
-        assert_eq!(
-            navigation_disposition(&desktop, "hosted", "https://lemma.work", ""),
-            NavigationDisposition::OpenExternal
-        );
-        assert_eq!(
-            new_window_disposition(&desktop, "hosted", "https://lemma.work", ""),
-            NewWindowDisposition::OpenExternal
-        );
-    }
-
-    /// A Windows user is never told about hardware they do not have.
-    ///
-    /// Both bundled pages ship in the Windows build. Every sentence about the
-    /// machine said "this Mac" -- including the recovery panel that names what
-    /// is about to be deleted, which is the worst possible place to describe
-    /// somebody else's computer.
-    ///
-    /// The three static splash strings used to be rewritten one element id at a
-    /// time, so the lines written from JS -- the boot subtitle, the ready
-    /// subtitle, the question the local-install screen asks -- were simply
-    /// missed. What is asserted here is the mechanism, not a list of strings:
-    /// a whole-document pass plus the two places text is produced after it.
-    /// A new "this Mac" anywhere is then covered without anyone remembering to
-    /// extend anything.
-    #[test]
-    fn every_page_that_ships_on_windows_renames_the_machine() {
-        let splash = include_str!("../ui/index.html").replace("\r\n", "\n");
-        let control = include_str!("../ui/control.js").replace("\r\n", "\n");
-
-        for (page, source) in [("index.html", &splash), ("control.js", &control)] {
-            assert!(
-                source.contains(r#"replace(/\bthis Mac\b/g, "this PC")"#),
-                "{page} has no device rewrite, so its copy is Mac-only",
-            );
-            assert!(
-                source.contains("NodeFilter.SHOW_TEXT"),
-                "{page} must rewrite the whole document, not a list of ids",
-            );
-            // Both pages carry inline or loaded script; rewriting a SCRIPT text
-            // node changes nothing anyone reads and leaves a DOM that no longer
-            // matches the file on disk.
-            assert!(
-                source.contains("SCRIPT|STYLE"),
-                "{page} rewrites script text as well as copy",
-            );
-        }
-
-        // The splash produces text after the document pass has run. Both
-        // producers have to go through the rewrite or the pass covers only the
-        // half of the copy that happens to be static.
-        assert!(
-            splash.contains("VOICE[key] = VOICE[key].map(forThisDevice)"),
-            "the phase table is written after the walk and needs its own pass",
-        );
-        assert!(
-            splash.contains("const text = forThisDevice(rawText);"),
-            "say() is the only writer of the serif line and must rewrite too",
-        );
-        // The settings window routes every error through one formatter.
-        assert!(
-            control.contains("return forThisDevice(match ? match[1]"),
-            "friendlyError is where the recovery copy is written",
-        );
-        // And the one sentence that is about the operating system rather than
-        // the box it runs on is named per platform, not rewritten.
-        assert!(control.contains(r#"IS_WINDOWS ? "Windows" : "macOS""#));
-    }
-
-    #[test]
-    fn local_settings_exposes_honest_runtime_repair_and_rollback_boundaries() {
-        let html = include_str!("../ui/control.html").replace("\r\n", "\n");
-        let script = include_str!("../ui/control.js").replace("\r\n", "\n");
-
-        assert!(html.contains("Signed release lifecycle"));
-        assert!(script.contains("repair_runtime"));
-        assert!(script.contains("open_developer_tools"));
-        assert!(html.contains("Developer tools"));
-        assert!(html.contains("id=\"network-contract\""));
-        assert!(html.contains("id=\"connector-callback\""));
-        assert!(script.contains("snapshot.state?.api_url"));
-        assert!(!html.contains("http://app.lemma.localhost:8711/api/v1/connectors"));
-        // The rollback notice used to be here, toggled `hidden = rollbackAvailable`
-        // against a value hardcoded `false` -- so it was *permanently* on
-        // screen, explaining a feature that does not exist. A standing
-        // paragraph about something that has never happened is a bug, not a
-        // flag, and the Previous runtime card now says what is actually true.
-        assert!(
-            !html.contains("Rollback stays unavailable"),
-            "a notice that can never be dismissed is not a boundary, it is noise",
-        );
-        assert!(html.contains("Reinstalling an earlier Lemma from the release page"));
-        assert!(html.contains("Databases, files, and workspaces are preserved"));
-    }
-
-    #[test]
-    fn cloudflare_sharing_defaults_to_safe_automatic_provisioning() {
-        let html = include_str!("../ui/control.html").replace("\r\n", "\n");
-        let script = include_str!("../ui/control.js").replace("\r\n", "\n");
-
-        assert!(html.contains("Automatic setup · recommended"));
-        assert!(html.contains("Use an existing named tunnel"));
-        assert!(
-            html.contains("Cloudflare automatic setup stores only its generated tunnel credential")
-        );
-        assert!(script.contains("payload.cloudflare_setup = $(\"cloudflare-setup\").value"));
-        assert!(script.contains("public_warning_confirmed: true"));
-        assert!(!script.contains("--overwrite-dns"));
-    }
-
-    #[test]
-    fn local_models_are_reached_through_a_provider_endpoint_not_an_app_owned_server() {
-        let html = include_str!("../ui/control.html").replace("\r\n", "\n");
-        let script = include_str!("../ui/control.js").replace("\r\n", "\n");
-
-        // Ollama and LM Studio are the supported local-model path: they are
-        // ordinary OpenAI-compatible endpoints the user already runs, so they
-        // only prefill the provider form and never give Lemma a model process
-        // of its own to install, supervise, or free memory for.
-        assert!(html.contains("data-preset=\"ollama\""));
-        assert!(html.contains("data-preset=\"lmstudio\""));
-        assert!(script.contains("http://127.0.0.1:11434/v1"));
-        assert!(script.contains("http://127.0.0.1:1234/v1"));
-        assert!(!script.contains("local_ai_action"));
-        assert!(!html.contains("data-page=\"models\""));
-    }
-
-    #[test]
-    fn the_default_model_is_chosen_from_the_providers_own_list() {
-        let html = include_str!("../ui/control.html").replace("\r\n", "\n");
-        let script = include_str!("../ui/control.js").replace("\r\n", "\n");
-
-        // Typing a model id from memory was the old contract and the reason a
-        // correct provider could still be applied with a model it does not
-        // serve. The default is now a <select> populated by the probe, and the
-        // free-text list it replaced must not come back.
-        assert!(html.contains("<select id=\"ai-model\">"));
-        assert!(!html.contains("id=\"ai-models\""));
-        // The probe answers the press that made it, rather than broadcasting to
-        // whichever page happens to be listening.
-        assert!(script.contains("const models = await invoke(\"discover_provider_models\""));
-        assert!(!script.contains("config.models"));
-    }
-
-    #[test]
-    fn local_settings_navigation_is_restricted_to_trusted_packaged_and_dev_assets() {
-        assert!(control_navigation_allowed(
-            &tauri::Url::parse(&native_assets::url("control.html", None)).unwrap()
-        ));
-        assert!(control_navigation_allowed(
-            &tauri::Url::parse("http://127.0.0.1:1430/control.html").unwrap()
-        ));
-        assert!(!control_navigation_allowed(
-            &tauri::Url::parse("http://127.0.0.1:1431/control.html").unwrap()
-        ));
-        assert!(!control_navigation_allowed(
-            &tauri::Url::parse("http://127.0.0.1:1430/index.html").unwrap()
-        ));
-        assert!(!control_navigation_allowed(
-            &tauri::Url::parse(&native_assets::url("index.html", None)).unwrap()
-        ));
-        assert!(!control_navigation_allowed(
-            &tauri::Url::parse("https://example.com/control.html").unwrap()
-        ));
-    }
-
-    #[test]
-    fn ready_auto_navigation_only_treats_the_native_installer_as_splash() {
-        assert!(native_splash_url(
-            &tauri::Url::parse(&native_assets::url("index.html", None)).unwrap()
-        ));
-        assert!(native_splash_url(
-            &tauri::Url::parse(&native_assets::url("", None)).unwrap()
-        ));
-        assert!(!native_splash_url(
-            &tauri::Url::parse("http://app.lemma.localhost:3711/").unwrap()
-        ));
-        assert!(!native_splash_url(
-            &tauri::Url::parse(&native_assets::url("control.html", None)).unwrap()
-        ));
-    }
-
-    #[test]
-    fn the_splash_is_reachable_in_the_build_we_are_running() {
-        // `tauri://localhost` does not exist under `cargo tauri dev` — the CLI
-        // serves frontendDist over loopback instead. Navigating there anyway
-        // left the window white on "index.html not found" until the workspace
-        // URL replaced it a minute later, so every startup stage the splash
-        // exists to show went unseen.
-        let splash = tauri::Url::parse(&native_asset_url("index.html")).unwrap();
-        assert!(
-            native_splash_url(&splash),
-            "the URL show_splash navigates to must be recognised as the splash: {splash}"
-        );
-        assert!(
-            trusted_native_asset_url(&splash),
-            "and must be a trusted native asset origin: {splash}"
-        );
-        if cfg!(debug_assertions) {
-            assert_eq!(splash.port(), Some(DEV_ASSET_PORT));
-        } else {
-            assert_eq!(
-                splash.scheme(),
-                if cfg!(windows) { "http" } else { "tauri" }
-            );
-        }
-    }
-
-    #[test]
-    fn the_splash_survives_the_navigation_gate_in_local_mode() {
-        // Local mode denies local http destinations that are not the workspace,
-        // which is what kept a dev build's splash off the screen: it is served
-        // over loopback http, so it looked exactly like the thing that rule
-        // exists to block.
-        let splash = tauri::Url::parse(&native_asset_url("index.html")).unwrap();
-        assert_eq!(
-            navigation_disposition(
-                &splash,
-                "local",
-                "http://app.lemma.localhost:52501",
-                "http://app.lemma.localhost:52502",
-            ),
-            NavigationDisposition::Allow,
-            "the splash must be allowed to load: {splash}"
-        );
-    }
-
-    #[test]
-    fn local_mode_still_denies_a_local_destination_that_is_not_ours() {
-        // The allowance above must not become a hole: an arbitrary loopback
-        // port is still refused in local mode.
-        assert_eq!(
-            navigation_disposition(
-                &tauri::Url::parse("http://127.0.0.1:9999/").unwrap(),
-                "local",
-                "http://app.lemma.localhost:52501",
-                "http://app.lemma.localhost:52502",
-            ),
-            NavigationDisposition::Deny
-        );
-    }
-
-    #[test]
-    fn the_dev_asset_origin_is_still_refused_for_anything_but_bundled_pages() {
-        // The development exception widens the trusted origin; it must not
-        // widen what a *remote* page can reach.
-        assert!(!trusted_native_asset_url(
-            &tauri::Url::parse("http://127.0.0.1:3000/index.html").unwrap()
-        ));
-        assert!(!trusted_control_url(
-            &tauri::Url::parse(&native_asset_url("index.html")).unwrap()
-        ));
-        assert!(trusted_control_url(
-            &tauri::Url::parse(&native_asset_url("control.html")).unwrap()
-        ));
-    }
-
-    #[test]
-    fn periodic_stopped_status_keeps_an_active_startup_phase_visible() {
-        assert!(should_preserve_inflight_phase(
-            "shell-start-123",
-            "infrastructure-health",
-            "stopped"
-        ));
-        assert!(!should_preserve_inflight_phase(
-            "",
-            "infrastructure-health",
-            "stopped"
-        ));
-        assert!(!should_preserve_inflight_phase(
-            "shell-stop-123",
-            "stopped",
-            "stopped"
-        ));
-    }
-
-    #[test]
-    fn desktop_config_replacement_never_exposes_a_partial_file() {
-        let root = tempfile::tempdir().unwrap();
-        let destination = root.path().join("desktop-config.json");
-        let source = root.path().join("desktop-config.json.next");
-        std::fs::write(&destination, br#"{"revision":1}"#).unwrap();
-        std::fs::write(&source, br#"{"revision":2}"#).unwrap();
-
-        config_store::replace(&source, &destination).unwrap();
-
-        assert_eq!(
-            std::fs::read_to_string(&destination).unwrap(),
-            r#"{"revision":2}"#
-        );
-        assert!(!source.exists());
-    }
-
-    #[test]
-    fn macos_allows_only_the_local_http_frontend_and_app_subdomains() {
-        let plist = include_str!("../Info.plist").replace("\r\n", "\n");
-
-        assert!(plist.contains("NSAllowsLocalNetworking"));
-        assert!(plist.contains("lemma.localhost"));
-        assert!(plist.contains("NSIncludesSubdomains"));
-        assert!(!plist.contains("NSAllowsArbitraryLoads"));
-        assert!(!plist.contains("NSAllowsArbitraryLoadsInWebContent"));
-    }
-
-    #[test]
-    fn legacy_connection_preferences_require_the_released_chooser_once() {
-        assert_eq!(
-            configured_connection_mode(&json!({"connectionMode": "hosted"})),
-            "undecided"
-        );
-        assert_eq!(
-            configured_connection_mode(&json!({
-                "connectionMode": "local",
-                "connectionModePromptRevision": CONNECTION_MODE_PROMPT_REVISION,
-            })),
-            "local"
-        );
-    }
-
-    #[test]
-    fn hosted_auth_uses_browser_handoff_while_local_auth_stays_in_app() {
-        assert_eq!(
-            desktop_auth_url("https://lemma.work", "signup"),
-            "https://lemma.work/auth/desktop?mode=signup"
-        );
-        assert_eq!(
-            local_auth_url("http://app.lemma.localhost:3711/", "signup"),
-            "http://app.lemma.localhost:3711/auth?show=signup"
-        );
-    }
-
-    #[test]
-    fn packaged_runtime_root_never_falls_back_to_the_build_checkout() {
-        let executable =
-            std::path::Path::new("/Applications/Lemma.app/Contents/MacOS/lemma-desktop");
-        let checkout = std::path::Path::new("/Users/developer/lemma-platform/desktop");
-
-        assert_eq!(
-            default_runtime_root(Some(executable), checkout, false),
-            std::path::Path::new("/Applications/Lemma.app/Contents/MacOS")
-        );
-        assert_eq!(
-            default_runtime_root(Some(executable), checkout, true),
-            std::path::Path::new("/Users/developer/lemma-platform")
-        );
-    }
-
-    #[test]
-    fn local_desktop_context_disables_email_verification_before_page_scripts() {
-        let local = desktop_context_script("local");
-        let hosted = desktop_context_script("hosted");
-
-        assert!(local.contains("AUTH_EMAIL_VERIFICATION_REQUIRED: \"false\""));
-        assert!(!hosted.contains("AUTH_EMAIL_VERIFICATION_REQUIRED"));
-    }
-
-    #[test]
-    fn native_material_attributes_survive_navigation() {
-        // Local mode navigates the main window from the splash to the
-        // workspace. Anything set by a one-shot eval is gone by then, so both
-        // attributes have to be written by the initialization script, which
-        // Tauri re-runs for every document.
-        let script = desktop_context_script("local");
-
-        assert!(script.contains("setAttribute('data-desktop-platform', d.platform)"));
-        assert!(script.contains("if (d.vibrancy) root.setAttribute('data-desktop-vibrancy'"));
-        assert!(script.contains("if (d.systemAccent) root.style.setProperty('--accent-rgb'"));
-    }
-}
+#[cfg(test)]
+mod tests;
